chore: move query into protected
