+++ apps/server/ent/internal/schema.go
@@ -6,4 +6,4 @@
 // Package internal holds a loadable version of the latest schema.
 package internal
 
-const Schema = "{\"Schema\":\"fijoy/ent/schema\",\"Package\":\"fijoy/ent\",\"Schemas\":[{\"name\":\"Account\",\"config\":{\"Table\":\"\"},\"edges\":[{\"name\":\"profile\",\"type\":\"Profile\",\"ref_name\":\"accounts\",\"unique\":true,\"inverse\":true,\"required\":true,\"annotations\":{\"EntGQL\":{\"Skip\":48}}},{\"name\":\"transaction_entries\",\"type\":\"TransactionEntry\"},{\"name\":\"snapshot_accounts\",\"type\":\"SnapshotAccount\"}],\"fields\":[{\"name\":\"create_time\",\"type\":{\"Type\":2,\"Ident\":\"\",\"PkgPath\":\"time\",\"PkgName\":\"\",\"Nillable\":false,\"RType\":null},\"default\":true,\"default_kind\":19,\"immutable\":true,\"position\":{\"Index\":0,\"MixedIn\":true,\"MixinIndex\":0}},{\"name\":\"update_time\",\"type\":{\"Type\":2,\"Ident\":\"\",\"PkgPath\":\"time\",\"PkgName\":\"\",\"Nillable\":false,\"RType\":null},\"default\":true,\"default_kind\":19,\"update_default\":true,\"position\":{\"Index\":1,\"MixedIn\":true,\"MixinIndex\":0}},{\"name\":\"name\",\"type\":{\"Type\":7,\"Ident\":\"\",\"PkgPath\":\"\",\"PkgName\":\"\",\"Nillable\":false,\"RType\":null},\"validators\":1,\"position\":{\"Index\":0,\"MixedIn\":false,\"MixinIndex\":0}},{\"name\":\"institution\",\"type\":{\"Type\":7,\"Ident\":\"\",\"PkgPath\":\"\",\"PkgName\":\"\",\"Nillable\":false,\"RType\":null},\"validators\":1,\"position\":{\"Index\":1,\"MixedIn\":false,\"MixinIndex\":0},\"comment\":\"The financial institution that holds this account\"},{\"name\":\"account_type\",\"type\":{\"Type\":6,\"Ident\":\"account.AccountType\",\"PkgPath\":\"\",\"PkgName\":\"\",\"Nillable\":false,\"RType\":null},\"enums\":[{\"N\":\"liquidity\",\"V\":\"liquidity\"},{\"N\":\"investment\",\"V\":\"investment\"},{\"N\":\"property\",\"V\":\"property\"},{\"N\":\"receivable\",\"V\":\"receivable\"},{\"N\":\"liability\",\"V\":\"liability\"}],\"immutable\":true,\"position\":{\"Index\":2,\"MixedIn\":false,\"MixinIndex\":0}},{\"name\":\"investment_type\",\"type\":{\"Type\":6,\"Ident\":\"account.InvestmentType\",\"PkgPath\":\"\",\"PkgName\":\"\",\"Nillable\":false,\"RType\":null},\"enums\":[{\"N\":\"non_investment\",\"V\":\"non_investment\"},{\"N\":\"taxable\",\"V\":\"taxable\"},{\"N\":\"rrsp\",\"V\":\"rrsp\"},{\"N\":\"tfsa\",\"V\":\"tfsa\"},{\"N\":\"fhsa\",\"V\":\"fhsa\"}],\"position\":{\"Index\":3,\"MixedIn\":false,\"MixinIndex\":0}},{\"name\":\"currency_code\",\"type\":{\"Type\":7,\"Ident\":\"\",\"PkgPath\":\"\",\"PkgName\":\"\",\"Nillable\":false,\"RType\":null},\"immutable\":true,\"validators\":1,\"position\":{\"Index\":4,\"MixedIn\":false,\"MixinIndex\":0}},{\"name\":\"ticker\",\"type\":{\"Type\":7,\"Ident\":\"\",\"PkgPath\":\"\",\"PkgName\":\"\",\"Nillable\":false,\"RType\":null},\"immutable\":true,\"validators\":1,\"position\":{\"Index\":5,\"MixedIn\":false,\"MixinIndex\":0}},{\"name\":\"ticker_type\",\"type\":{\"Type\":6,\"Ident\":\"account.TickerType\",\"PkgPath\":\"\",\"PkgName\":\"\",\"Nillable\":false,\"RType\":null},\"enums\":[{\"N\":\"currency\",\"V\":\"currency\"},{\"N\":\"stock\",\"V\":\"stock\"},{\"N\":\"crypto\",\"V\":\"crypto\"}],\"immutable\":true,\"position\":{\"Index\":6,\"MixedIn\":false,\"MixinIndex\":0}},{\"name\":\"amount\",\"type\":{\"Type\":20,\"Ident\":\"decimal.Decimal\",\"PkgPath\":\"github.com/shopspring/decimal\",\"PkgName\":\"decimal\",\"Nillable\":false,\"RType\":{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":{\"Abs\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Add\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Atan\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"BigFloat\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"*big.Float\",\"Kind\":22,\"PkgPath\":\"\",\"Methods\":null}]},\"BigInt\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"*big.Int\",\"Kind\":22,\"PkgPath\":\"\",\"Methods\":null}]},\"Ceil\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Cmp\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"int\",\"Ident\":\"int\",\"Kind\":2,\"PkgPath\":\"\",\"Methods\":null}]},\"Coefficient\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"*big.Int\",\"Kind\":22,\"PkgPath\":\"\",\"Methods\":null}]},\"CoefficientInt64\":{\"In\":[],\"Out\":[{\"Name\":\"int64\",\"Ident\":\"int64\",\"Kind\":6,\"PkgPath\":\"\",\"Methods\":null}]},\"Compare\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"int\",\"Ident\":\"int\",\"Kind\":2,\"PkgPath\":\"\",\"Methods\":null}]},\"Copy\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Cos\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Div\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"DivRound\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Equal\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"Equals\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"ExpHullAbrham\":{\"In\":[{\"Name\":\"uint32\",\"Ident\":\"uint32\",\"Kind\":10,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"ExpTaylor\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"Exponent\":{\"In\":[],\"Out\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}]},\"Float64\":{\"In\":[],\"Out\":[{\"Name\":\"float64\",\"Ident\":\"float64\",\"Kind\":14,\"PkgPath\":\"\",\"Methods\":null},{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"Floor\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"GobDecode\":{\"In\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"GobEncode\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"GreaterThan\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"GreaterThanOrEqual\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"InexactFloat64\":{\"In\":[],\"Out\":[{\"Name\":\"float64\",\"Ident\":\"float64\",\"Kind\":14,\"PkgPath\":\"\",\"Methods\":null}]},\"IntPart\":{\"In\":[],\"Out\":[{\"Name\":\"int64\",\"Ident\":\"int64\",\"Kind\":6,\"PkgPath\":\"\",\"Methods\":null}]},\"IsInteger\":{\"In\":[],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"IsNegative\":{\"In\":[],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"IsPositive\":{\"In\":[],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"IsZero\":{\"In\":[],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"LessThan\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"LessThanOrEqual\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"Ln\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"MarshalBinary\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"MarshalJSON\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"MarshalText\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"Mod\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Mul\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Neg\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"NumDigits\":{\"In\":[],\"Out\":[{\"Name\":\"int\",\"Ident\":\"int\",\"Kind\":2,\"PkgPath\":\"\",\"Methods\":null}]},\"Pow\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"PowBigInt\":{\"In\":[{\"Name\":\"\",\"Ident\":\"*big.Int\",\"Kind\":22,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"PowInt32\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"PowWithPrecision\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"QuoRem\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Rat\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"*big.Rat\",\"Kind\":22,\"PkgPath\":\"\",\"Methods\":null}]},\"Round\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"RoundBank\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"RoundCash\":{\"In\":[{\"Name\":\"uint8\",\"Ident\":\"uint8\",\"Kind\":8,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"RoundCeil\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"RoundDown\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"RoundFloor\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"RoundUp\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Scan\":{\"In\":[{\"Name\":\"\",\"Ident\":\"interface {}\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"Shift\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Sign\":{\"In\":[],\"Out\":[{\"Name\":\"int\",\"Ident\":\"int\",\"Kind\":2,\"PkgPath\":\"\",\"Methods\":null}]},\"Sin\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"String\":{\"In\":[],\"Out\":[{\"Name\":\"string\",\"Ident\":\"string\",\"Kind\":24,\"PkgPath\":\"\",\"Methods\":null}]},\"StringFixed\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"string\",\"Ident\":\"string\",\"Kind\":24,\"PkgPath\":\"\",\"Methods\":null}]},\"StringFixedBank\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"string\",\"Ident\":\"string\",\"Kind\":24,\"PkgPath\":\"\",\"Methods\":null}]},\"StringFixedCash\":{\"In\":[{\"Name\":\"uint8\",\"Ident\":\"uint8\",\"Kind\":8,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"string\",\"Ident\":\"string\",\"Kind\":24,\"PkgPath\":\"\",\"Methods\":null}]},\"StringScaled\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"string\",\"Ident\":\"string\",\"Kind\":24,\"PkgPath\":\"\",\"Methods\":null}]},\"Sub\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Tan\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Truncate\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"UnmarshalBinary\":{\"In\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"UnmarshalJSON\":{\"In\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"UnmarshalText\":{\"In\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"Value\":{\"In\":[],\"Out\":[{\"Name\":\"Value\",\"Ident\":\"driver.Value\",\"Kind\":20,\"PkgPath\":\"database/sql/driver\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]}}}},\"position\":{\"Index\":7,\"MixedIn\":false,\"MixinIndex\":0},\"schema_type\":{\"mysql\":\"decimal(36,18)\",\"postgres\":\"numeric(36,18)\"},\"annotations\":{\"EntGQL\":{\"Type\":\"String\"}},\"comment\":\"The unit amount of share or money in this account\"},{\"name\":\"value\",\"type\":{\"Type\":20,\"Ident\":\"decimal.Decimal\",\"PkgPath\":\"github.com/shopspring/decimal\",\"PkgName\":\"decimal\",\"Nillable\":false,\"RType\":{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":{\"Abs\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Add\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Atan\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"BigFloat\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"*big.Float\",\"Kind\":22,\"PkgPath\":\"\",\"Methods\":null}]},\"BigInt\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"*big.Int\",\"Kind\":22,\"PkgPath\":\"\",\"Methods\":null}]},\"Ceil\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Cmp\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"int\",\"Ident\":\"int\",\"Kind\":2,\"PkgPath\":\"\",\"Methods\":null}]},\"Coefficient\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"*big.Int\",\"Kind\":22,\"PkgPath\":\"\",\"Methods\":null}]},\"CoefficientInt64\":{\"In\":[],\"Out\":[{\"Name\":\"int64\",\"Ident\":\"int64\",\"Kind\":6,\"PkgPath\":\"\",\"Methods\":null}]},\"Compare\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"int\",\"Ident\":\"int\",\"Kind\":2,\"PkgPath\":\"\",\"Methods\":null}]},\"Copy\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Cos\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Div\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"DivRound\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Equal\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"Equals\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"ExpHullAbrham\":{\"In\":[{\"Name\":\"uint32\",\"Ident\":\"uint32\",\"Kind\":10,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"ExpTaylor\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"Exponent\":{\"In\":[],\"Out\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}]},\"Float64\":{\"In\":[],\"Out\":[{\"Name\":\"float64\",\"Ident\":\"float64\",\"Kind\":14,\"PkgPath\":\"\",\"Methods\":null},{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"Floor\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"GobDecode\":{\"In\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"GobEncode\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"GreaterThan\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"GreaterThanOrEqual\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"InexactFloat64\":{\"In\":[],\"Out\":[{\"Name\":\"float64\",\"Ident\":\"float64\",\"Kind\":14,\"PkgPath\":\"\",\"Methods\":null}]},\"IntPart\":{\"In\":[],\"Out\":[{\"Name\":\"int64\",\"Ident\":\"int64\",\"Kind\":6,\"PkgPath\":\"\",\"Methods\":null}]},\"IsInteger\":{\"In\":[],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"IsNegative\":{\"In\":[],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"IsPositive\":{\"In\":[],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"IsZero\":{\"In\":[],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"LessThan\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"LessThanOrEqual\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"Ln\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"MarshalBinary\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"MarshalJSON\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"MarshalText\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"Mod\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Mul\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Neg\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"NumDigits\":{\"In\":[],\"Out\":[{\"Name\":\"int\",\"Ident\":\"int\",\"Kind\":2,\"PkgPath\":\"\",\"Methods\":null}]},\"Pow\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"PowBigInt\":{\"In\":[{\"Name\":\"\",\"Ident\":\"*big.Int\",\"Kind\":22,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"PowInt32\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"PowWithPrecision\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"QuoRem\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Rat\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"*big.Rat\",\"Kind\":22,\"PkgPath\":\"\",\"Methods\":null}]},\"Round\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"RoundBank\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"RoundCash\":{\"In\":[{\"Name\":\"uint8\",\"Ident\":\"uint8\",\"Kind\":8,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"RoundCeil\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"RoundDown\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"RoundFloor\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"RoundUp\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Scan\":{\"In\":[{\"Name\":\"\",\"Ident\":\"interface {}\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"Shift\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Sign\":{\"In\":[],\"Out\":[{\"Name\":\"int\",\"Ident\":\"int\",\"Kind\":2,\"PkgPath\":\"\",\"Methods\":null}]},\"Sin\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"String\":{\"In\":[],\"Out\":[{\"Name\":\"string\",\"Ident\":\"string\",\"Kind\":24,\"PkgPath\":\"\",\"Methods\":null}]},\"StringFixed\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"string\",\"Ident\":\"string\",\"Kind\":24,\"PkgPath\":\"\",\"Methods\":null}]},\"StringFixedBank\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"string\",\"Ident\":\"string\",\"Kind\":24,\"PkgPath\":\"\",\"Methods\":null}]},\"StringFixedCash\":{\"In\":[{\"Name\":\"uint8\",\"Ident\":\"uint8\",\"Kind\":8,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"string\",\"Ident\":\"string\",\"Kind\":24,\"PkgPath\":\"\",\"Methods\":null}]},\"StringScaled\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"string\",\"Ident\":\"string\",\"Kind\":24,\"PkgPath\":\"\",\"Methods\":null}]},\"Sub\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Tan\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Truncate\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"UnmarshalBinary\":{\"In\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"UnmarshalJSON\":{\"In\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"UnmarshalText\":{\"In\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"Value\":{\"In\":[],\"Out\":[{\"Name\":\"Value\",\"Ident\":\"driver.Value\",\"Kind\":20,\"PkgPath\":\"database/sql/driver\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]}}}},\"position\":{\"Index\":8,\"MixedIn\":false,\"MixinIndex\":0},\"schema_type\":{\"mysql\":\"decimal(18,10)\",\"postgres\":\"numeric(18,10)\"},\"annotations\":{\"EntGQL\":{\"Skip\":48,\"Type\":\"String\"}},\"comment\":\"The value of 1 share in the native currency. If this is just a currency account, then this field will be 1\"},{\"name\":\"balance\",\"type\":{\"Type\":20,\"Ident\":\"decimal.Decimal\",\"PkgPath\":\"github.com/shopspring/decimal\",\"PkgName\":\"decimal\",\"Nillable\":false,\"RType\":{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":{\"Abs\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Add\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Atan\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"BigFloat\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"*big.Float\",\"Kind\":22,\"PkgPath\":\"\",\"Methods\":null}]},\"BigInt\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"*big.Int\",\"Kind\":22,\"PkgPath\":\"\",\"Methods\":null}]},\"Ceil\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Cmp\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"int\",\"Ident\":\"int\",\"Kind\":2,\"PkgPath\":\"\",\"Methods\":null}]},\"Coefficient\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"*big.Int\",\"Kind\":22,\"PkgPath\":\"\",\"Methods\":null}]},\"CoefficientInt64\":{\"In\":[],\"Out\":[{\"Name\":\"int64\",\"Ident\":\"int64\",\"Kind\":6,\"PkgPath\":\"\",\"Methods\":null}]},\"Compare\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"int\",\"Ident\":\"int\",\"Kind\":2,\"PkgPath\":\"\",\"Methods\":null}]},\"Copy\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Cos\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Div\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"DivRound\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Equal\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"Equals\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"ExpHullAbrham\":{\"In\":[{\"Name\":\"uint32\",\"Ident\":\"uint32\",\"Kind\":10,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"ExpTaylor\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"Exponent\":{\"In\":[],\"Out\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}]},\"Float64\":{\"In\":[],\"Out\":[{\"Name\":\"float64\",\"Ident\":\"float64\",\"Kind\":14,\"PkgPath\":\"\",\"Methods\":null},{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"Floor\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"GobDecode\":{\"In\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"GobEncode\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"GreaterThan\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"GreaterThanOrEqual\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"InexactFloat64\":{\"In\":[],\"Out\":[{\"Name\":\"float64\",\"Ident\":\"float64\",\"Kind\":14,\"PkgPath\":\"\",\"Methods\":null}]},\"IntPart\":{\"In\":[],\"Out\":[{\"Name\":\"int64\",\"Ident\":\"int64\",\"Kind\":6,\"PkgPath\":\"\",\"Methods\":null}]},\"IsInteger\":{\"In\":[],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"IsNegative\":{\"In\":[],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"IsPositive\":{\"In\":[],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"IsZero\":{\"In\":[],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"LessThan\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"LessThanOrEqual\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"Ln\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"MarshalBinary\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"MarshalJSON\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"MarshalText\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"Mod\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Mul\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Neg\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"NumDigits\":{\"In\":[],\"Out\":[{\"Name\":\"int\",\"Ident\":\"int\",\"Kind\":2,\"PkgPath\":\"\",\"Methods\":null}]},\"Pow\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"PowBigInt\":{\"In\":[{\"Name\":\"\",\"Ident\":\"*big.Int\",\"Kind\":22,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"PowInt32\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"PowWithPrecision\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"QuoRem\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Rat\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"*big.Rat\",\"Kind\":22,\"PkgPath\":\"\",\"Methods\":null}]},\"Round\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"RoundBank\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"RoundCash\":{\"In\":[{\"Name\":\"uint8\",\"Ident\":\"uint8\",\"Kind\":8,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"RoundCeil\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"RoundDown\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"RoundFloor\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"RoundUp\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Scan\":{\"In\":[{\"Name\":\"\",\"Ident\":\"interface {}\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"Shift\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Sign\":{\"In\":[],\"Out\":[{\"Name\":\"int\",\"Ident\":\"int\",\"Kind\":2,\"PkgPath\":\"\",\"Methods\":null}]},\"Sin\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"String\":{\"In\":[],\"Out\":[{\"Name\":\"string\",\"Ident\":\"string\",\"Kind\":24,\"PkgPath\":\"\",\"Methods\":null}]},\"StringFixed\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"string\",\"Ident\":\"string\",\"Kind\":24,\"PkgPath\":\"\",\"Methods\":null}]},\"StringFixedBank\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"string\",\"Ident\":\"string\",\"Kind\":24,\"PkgPath\":\"\",\"Methods\":null}]},\"StringFixedCash\":{\"In\":[{\"Name\":\"uint8\",\"Ident\":\"uint8\",\"Kind\":8,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"string\",\"Ident\":\"string\",\"Kind\":24,\"PkgPath\":\"\",\"Methods\":null}]},\"StringScaled\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"string\",\"Ident\":\"string\",\"Kind\":24,\"PkgPath\":\"\",\"Methods\":null}]},\"Sub\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Tan\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Truncate\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"UnmarshalBinary\":{\"In\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"UnmarshalJSON\":{\"In\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"UnmarshalText\":{\"In\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"Value\":{\"In\":[],\"Out\":[{\"Name\":\"Value\",\"Ident\":\"driver.Value\",\"Kind\":20,\"PkgPath\":\"database/sql/driver\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]}}}},\"position\":{\"Index\":9,\"MixedIn\":false,\"MixinIndex\":0},\"schema_type\":{\"mysql\":\"decimal(36,18)\",\"postgres\":\"numeric(36,18)\"},\"annotations\":{\"EntGQL\":{\"Skip\":48,\"Type\":\"String\"}},\"comment\":\"The total balance of this account in its currency\"},{\"name\":\"archived\",\"type\":{\"Type\":1,\"Ident\":\"\",\"PkgPath\":\"\",\"PkgName\":\"\",\"Nillable\":false,\"RType\":null},\"default\":true,\"default_value\":false,\"default_kind\":1,\"position\":{\"Index\":10,\"MixedIn\":false,\"MixinIndex\":0}}],\"annotations\":{\"EntGQL\":{\"MutationInputs\":[{\"IsCreate\":true},{}],\"QueryField\":{},\"RelayConnection\":true},\"EntSQL\":{\"increment_start\":0}}},{\"name\":\"Category\",\"config\":{\"Table\":\"\"},\"edges\":[{\"name\":\"profile\",\"type\":\"Profile\",\"ref_name\":\"categories\",\"unique\":true,\"inverse\":true,\"required\":true,\"annotations\":{\"EntGQL\":{\"Skip\":48}}},{\"name\":\"transactions\",\"type\":\"Transaction\"}],\"fields\":[{\"name\":\"create_time\",\"type\":{\"Type\":2,\"Ident\":\"\",\"PkgPath\":\"time\",\"PkgName\":\"\",\"Nillable\":false,\"RType\":null},\"default\":true,\"default_kind\":19,\"immutable\":true,\"position\":{\"Index\":0,\"MixedIn\":true,\"MixinIndex\":0}},{\"name\":\"update_time\",\"type\":{\"Type\":2,\"Ident\":\"\",\"PkgPath\":\"time\",\"PkgName\":\"\",\"Nillable\":false,\"RType\":null},\"default\":true,\"default_kind\":19,\"update_default\":true,\"position\":{\"Index\":1,\"MixedIn\":true,\"MixinIndex\":0}},{\"name\":\"name\",\"type\":{\"Type\":7,\"Ident\":\"\",\"PkgPath\":\"\",\"PkgName\":\"\",\"Nillable\":false,\"RType\":null},\"validators\":1,\"position\":{\"Index\":0,\"MixedIn\":false,\"MixinIndex\":0}},{\"name\":\"category_type\",\"type\":{\"Type\":6,\"Ident\":\"category.CategoryType\",\"PkgPath\":\"\",\"PkgName\":\"\",\"Nillable\":false,\"RType\":null},\"enums\":[{\"N\":\"expense\",\"V\":\"expense\"},{\"N\":\"income\",\"V\":\"income\"},{\"N\":\"transfer\",\"V\":\"transfer\"},{\"N\":\"sync\",\"V\":\"sync\"},{\"N\":\"init\",\"V\":\"init\"}],\"immutable\":true,\"position\":{\"Index\":1,\"MixedIn\":false,\"MixinIndex\":0}}],\"annotations\":{\"EntGQL\":{\"MutationInputs\":[{\"IsCreate\":true},{}],\"QueryField\":{}},\"EntSQL\":{\"increment_start\":38654705664}}},{\"name\":\"Profile\",\"config\":{\"Table\":\"\"},\"edges\":[{\"name\":\"users\",\"type\":\"User\",\"ref_name\":\"profiles\",\"through\":{\"N\":\"user_profiles\",\"T\":\"UserProfile\"},\"inverse\":true,\"required\":true,\"annotations\":{\"EntGQL\":{\"Skip\":48}}},{\"name\":\"accounts\",\"type\":\"Account\",\"annotations\":{\"EntSQL\":{\"on_delete\":\"CASCADE\"}}},{\"name\":\"transactions\",\"type\":\"Transaction\",\"annotations\":{\"EntSQL\":{\"on_delete\":\"CASCADE\"}}},{\"name\":\"snapshots\",\"type\":\"Snapshot\",\"annotations\":{\"EntSQL\":{\"on_delete\":\"CASCADE\"}}},{\"name\":\"categories\",\"type\":\"Category\",\"annotations\":{\"EntSQL\":{\"on_delete\":\"CASCADE\"}}}],\"fields\":[{\"name\":\"create_time\",\"type\":{\"Type\":2,\"Ident\":\"\",\"PkgPath\":\"time\",\"PkgName\":\"\",\"Nillable\":false,\"RType\":null},\"default\":true,\"default_kind\":19,\"immutable\":true,\"position\":{\"Index\":0,\"MixedIn\":true,\"MixinIndex\":0}},{\"name\":\"update_time\",\"type\":{\"Type\":2,\"Ident\":\"\",\"PkgPath\":\"time\",\"PkgName\":\"\",\"Nillable\":false,\"RType\":null},\"default\":true,\"default_kind\":19,\"update_default\":true,\"position\":{\"Index\":1,\"MixedIn\":true,\"MixinIndex\":0}},{\"name\":\"id\",\"type\":{\"Type\":12,\"Ident\":\"\",\"PkgPath\":\"\",\"PkgName\":\"\",\"Nillable\":false,\"RType\":null},\"position\":{\"Index\":0,\"MixedIn\":false,\"MixinIndex\":0}},{\"name\":\"name\",\"type\":{\"Type\":7,\"Ident\":\"\",\"PkgPath\":\"\",\"PkgName\":\"\",\"Nillable\":false,\"RType\":null},\"position\":{\"Index\":1,\"MixedIn\":false,\"MixinIndex\":0}},{\"name\":\"locale\",\"type\":{\"Type\":7,\"Ident\":\"\",\"PkgPath\":\"\",\"PkgName\":\"\",\"Nillable\":false,\"RType\":null},\"position\":{\"Index\":2,\"MixedIn\":false,\"MixinIndex\":0},\"annotations\":{\"EntGQL\":{\"Skip\":48}}},{\"name\":\"currencies\",\"type\":{\"Type\":3,\"Ident\":\"[]string\",\"PkgPath\":\"\",\"PkgName\":\"\",\"Nillable\":true,\"RType\":{\"Name\":\"\",\"Ident\":\"[]string\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":{}}},\"position\":{\"Index\":3,\"MixedIn\":false,\"MixinIndex\":0}},{\"name\":\"net_worth_goal\",\"type\":{\"Type\":20,\"Ident\":\"decimal.Decimal\",\"PkgPath\":\"github.com/shopspring/decimal\",\"PkgName\":\"decimal\",\"Nillable\":false,\"RType\":{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":{\"Abs\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Add\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Atan\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"BigFloat\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"*big.Float\",\"Kind\":22,\"PkgPath\":\"\",\"Methods\":null}]},\"BigInt\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"*big.Int\",\"Kind\":22,\"PkgPath\":\"\",\"Methods\":null}]},\"Ceil\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Cmp\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"int\",\"Ident\":\"int\",\"Kind\":2,\"PkgPath\":\"\",\"Methods\":null}]},\"Coefficient\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"*big.Int\",\"Kind\":22,\"PkgPath\":\"\",\"Methods\":null}]},\"CoefficientInt64\":{\"In\":[],\"Out\":[{\"Name\":\"int64\",\"Ident\":\"int64\",\"Kind\":6,\"PkgPath\":\"\",\"Methods\":null}]},\"Compare\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"int\",\"Ident\":\"int\",\"Kind\":2,\"PkgPath\":\"\",\"Methods\":null}]},\"Copy\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Cos\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Div\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"DivRound\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Equal\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"Equals\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"ExpHullAbrham\":{\"In\":[{\"Name\":\"uint32\",\"Ident\":\"uint32\",\"Kind\":10,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"ExpTaylor\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"Exponent\":{\"In\":[],\"Out\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}]},\"Float64\":{\"In\":[],\"Out\":[{\"Name\":\"float64\",\"Ident\":\"float64\",\"Kind\":14,\"PkgPath\":\"\",\"Methods\":null},{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"Floor\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"GobDecode\":{\"In\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"GobEncode\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"GreaterThan\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"GreaterThanOrEqual\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"InexactFloat64\":{\"In\":[],\"Out\":[{\"Name\":\"float64\",\"Ident\":\"float64\",\"Kind\":14,\"PkgPath\":\"\",\"Methods\":null}]},\"IntPart\":{\"In\":[],\"Out\":[{\"Name\":\"int64\",\"Ident\":\"int64\",\"Kind\":6,\"PkgPath\":\"\",\"Methods\":null}]},\"IsInteger\":{\"In\":[],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"IsNegative\":{\"In\":[],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"IsPositive\":{\"In\":[],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"IsZero\":{\"In\":[],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"LessThan\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"LessThanOrEqual\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"Ln\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"MarshalBinary\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"MarshalJSON\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"MarshalText\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"Mod\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Mul\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Neg\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"NumDigits\":{\"In\":[],\"Out\":[{\"Name\":\"int\",\"Ident\":\"int\",\"Kind\":2,\"PkgPath\":\"\",\"Methods\":null}]},\"Pow\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"PowBigInt\":{\"In\":[{\"Name\":\"\",\"Ident\":\"*big.Int\",\"Kind\":22,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"PowInt32\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"PowWithPrecision\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"QuoRem\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Rat\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"*big.Rat\",\"Kind\":22,\"PkgPath\":\"\",\"Methods\":null}]},\"Round\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"RoundBank\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"RoundCash\":{\"In\":[{\"Name\":\"uint8\",\"Ident\":\"uint8\",\"Kind\":8,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"RoundCeil\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"RoundDown\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"RoundFloor\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"RoundUp\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Scan\":{\"In\":[{\"Name\":\"\",\"Ident\":\"interface {}\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"Shift\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Sign\":{\"In\":[],\"Out\":[{\"Name\":\"int\",\"Ident\":\"int\",\"Kind\":2,\"PkgPath\":\"\",\"Methods\":null}]},\"Sin\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"String\":{\"In\":[],\"Out\":[{\"Name\":\"string\",\"Ident\":\"string\",\"Kind\":24,\"PkgPath\":\"\",\"Methods\":null}]},\"StringFixed\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"string\",\"Ident\":\"string\",\"Kind\":24,\"PkgPath\":\"\",\"Methods\":null}]},\"StringFixedBank\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"string\",\"Ident\":\"string\",\"Kind\":24,\"PkgPath\":\"\",\"Methods\":null}]},\"StringFixedCash\":{\"In\":[{\"Name\":\"uint8\",\"Ident\":\"uint8\",\"Kind\":8,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"string\",\"Ident\":\"string\",\"Kind\":24,\"PkgPath\":\"\",\"Methods\":null}]},\"StringScaled\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"string\",\"Ident\":\"string\",\"Kind\":24,\"PkgPath\":\"\",\"Methods\":null}]},\"Sub\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Tan\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Truncate\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"UnmarshalBinary\":{\"In\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"UnmarshalJSON\":{\"In\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"UnmarshalText\":{\"In\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"Value\":{\"In\":[],\"Out\":[{\"Name\":\"Value\",\"Ident\":\"driver.Value\",\"Kind\":20,\"PkgPath\":\"database/sql/driver\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]}}}},\"position\":{\"Index\":4,\"MixedIn\":false,\"MixinIndex\":0},\"schema_type\":{\"mysql\":\"decimal(36,18)\",\"postgres\":\"numeric(36,18)\"},\"annotations\":{\"EntGQL\":{\"Type\":\"String\"}}}],\"annotations\":{\"EntGQL\":{\"MutationInputs\":[{\"IsCreate\":true},{}],\"QueryField\":{}},\"EntSQL\":{\"increment_start\":4294967296}}},{\"name\":\"Snapshot\",\"config\":{\"Table\":\"\"},\"edges\":[{\"name\":\"profile\",\"type\":\"Profile\",\"ref_name\":\"snapshots\",\"unique\":true,\"inverse\":true,\"required\":true,\"annotations\":{\"EntGQL\":{\"Skip\":48}}},{\"name\":\"snapshot_accounts\",\"type\":\"SnapshotAccount\",\"annotations\":{\"EntSQL\":{\"on_delete\":\"CASCADE\"}}},{\"name\":\"snapshot_fx_rates\",\"type\":\"SnapshotFXRate\",\"annotations\":{\"EntSQL\":{\"on_delete\":\"CASCADE\"}}}],\"fields\":[{\"name\":\"create_time\",\"type\":{\"Type\":2,\"Ident\":\"\",\"PkgPath\":\"time\",\"PkgName\":\"\",\"Nillable\":false,\"RType\":null},\"default\":true,\"default_kind\":19,\"immutable\":true,\"position\":{\"Index\":0,\"MixedIn\":true,\"MixinIndex\":0}},{\"name\":\"update_time\",\"type\":{\"Type\":2,\"Ident\":\"\",\"PkgPath\":\"time\",\"PkgName\":\"\",\"Nillable\":false,\"RType\":null},\"default\":true,\"default_kind\":19,\"update_default\":true,\"position\":{\"Index\":1,\"MixedIn\":true,\"MixinIndex\":0}},{\"name\":\"note\",\"type\":{\"Type\":7,\"Ident\":\"\",\"PkgPath\":\"\",\"PkgName\":\"\",\"Nillable\":false,\"RType\":null},\"size\":2147483647,\"optional\":true,\"position\":{\"Index\":0,\"MixedIn\":false,\"MixinIndex\":0}}],\"annotations\":{\"EntGQL\":{\"QueryField\":{}},\"EntSQL\":{\"increment_start\":25769803776}}},{\"name\":\"SnapshotAccount\",\"config\":{\"Table\":\"\"},\"edges\":[{\"name\":\"account\",\"type\":\"Account\",\"ref_name\":\"snapshot_accounts\",\"unique\":true,\"inverse\":true,\"required\":true,\"immutable\":true,\"annotations\":{\"EntGQL\":{\"Skip\":48}}},{\"name\":\"snapshot\",\"type\":\"Snapshot\",\"ref_name\":\"snapshot_accounts\",\"unique\":true,\"inverse\":true,\"required\":true,\"immutable\":true,\"annotations\":{\"EntGQL\":{\"Skip\":48}}}],\"fields\":[{\"name\":\"create_time\",\"type\":{\"Type\":2,\"Ident\":\"\",\"PkgPath\":\"time\",\"PkgName\":\"\",\"Nillable\":false,\"RType\":null},\"default\":true,\"default_kind\":19,\"immutable\":true,\"position\":{\"Index\":0,\"MixedIn\":true,\"MixinIndex\":0}},{\"name\":\"update_time\",\"type\":{\"Type\":2,\"Ident\":\"\",\"PkgPath\":\"time\",\"PkgName\":\"\",\"Nillable\":false,\"RType\":null},\"default\":true,\"default_kind\":19,\"update_default\":true,\"position\":{\"Index\":1,\"MixedIn\":true,\"MixinIndex\":0}},{\"name\":\"amount\",\"type\":{\"Type\":20,\"Ident\":\"decimal.Decimal\",\"PkgPath\":\"github.com/shopspring/decimal\",\"PkgName\":\"decimal\",\"Nillable\":false,\"RType\":{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":{\"Abs\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Add\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Atan\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"BigFloat\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"*big.Float\",\"Kind\":22,\"PkgPath\":\"\",\"Methods\":null}]},\"BigInt\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"*big.Int\",\"Kind\":22,\"PkgPath\":\"\",\"Methods\":null}]},\"Ceil\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Cmp\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"int\",\"Ident\":\"int\",\"Kind\":2,\"PkgPath\":\"\",\"Methods\":null}]},\"Coefficient\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"*big.Int\",\"Kind\":22,\"PkgPath\":\"\",\"Methods\":null}]},\"CoefficientInt64\":{\"In\":[],\"Out\":[{\"Name\":\"int64\",\"Ident\":\"int64\",\"Kind\":6,\"PkgPath\":\"\",\"Methods\":null}]},\"Compare\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"int\",\"Ident\":\"int\",\"Kind\":2,\"PkgPath\":\"\",\"Methods\":null}]},\"Copy\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Cos\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Div\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"DivRound\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Equal\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"Equals\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"ExpHullAbrham\":{\"In\":[{\"Name\":\"uint32\",\"Ident\":\"uint32\",\"Kind\":10,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"ExpTaylor\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"Exponent\":{\"In\":[],\"Out\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}]},\"Float64\":{\"In\":[],\"Out\":[{\"Name\":\"float64\",\"Ident\":\"float64\",\"Kind\":14,\"PkgPath\":\"\",\"Methods\":null},{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"Floor\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"GobDecode\":{\"In\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"GobEncode\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"GreaterThan\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"GreaterThanOrEqual\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"InexactFloat64\":{\"In\":[],\"Out\":[{\"Name\":\"float64\",\"Ident\":\"float64\",\"Kind\":14,\"PkgPath\":\"\",\"Methods\":null}]},\"IntPart\":{\"In\":[],\"Out\":[{\"Name\":\"int64\",\"Ident\":\"int64\",\"Kind\":6,\"PkgPath\":\"\",\"Methods\":null}]},\"IsInteger\":{\"In\":[],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"IsNegative\":{\"In\":[],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"IsPositive\":{\"In\":[],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"IsZero\":{\"In\":[],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"LessThan\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"LessThanOrEqual\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"Ln\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"MarshalBinary\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"MarshalJSON\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"MarshalText\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"Mod\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Mul\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Neg\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"NumDigits\":{\"In\":[],\"Out\":[{\"Name\":\"int\",\"Ident\":\"int\",\"Kind\":2,\"PkgPath\":\"\",\"Methods\":null}]},\"Pow\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"PowBigInt\":{\"In\":[{\"Name\":\"\",\"Ident\":\"*big.Int\",\"Kind\":22,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"PowInt32\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"PowWithPrecision\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"QuoRem\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Rat\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"*big.Rat\",\"Kind\":22,\"PkgPath\":\"\",\"Methods\":null}]},\"Round\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"RoundBank\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"RoundCash\":{\"In\":[{\"Name\":\"uint8\",\"Ident\":\"uint8\",\"Kind\":8,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"RoundCeil\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"RoundDown\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"RoundFloor\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"RoundUp\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Scan\":{\"In\":[{\"Name\":\"\",\"Ident\":\"interface {}\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"Shift\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Sign\":{\"In\":[],\"Out\":[{\"Name\":\"int\",\"Ident\":\"int\",\"Kind\":2,\"PkgPath\":\"\",\"Methods\":null}]},\"Sin\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"String\":{\"In\":[],\"Out\":[{\"Name\":\"string\",\"Ident\":\"string\",\"Kind\":24,\"PkgPath\":\"\",\"Methods\":null}]},\"StringFixed\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"string\",\"Ident\":\"string\",\"Kind\":24,\"PkgPath\":\"\",\"Methods\":null}]},\"StringFixedBank\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"string\",\"Ident\":\"string\",\"Kind\":24,\"PkgPath\":\"\",\"Methods\":null}]},\"StringFixedCash\":{\"In\":[{\"Name\":\"uint8\",\"Ident\":\"uint8\",\"Kind\":8,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"string\",\"Ident\":\"string\",\"Kind\":24,\"PkgPath\":\"\",\"Methods\":null}]},\"StringScaled\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"string\",\"Ident\":\"string\",\"Kind\":24,\"PkgPath\":\"\",\"Methods\":null}]},\"Sub\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Tan\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Truncate\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"UnmarshalBinary\":{\"In\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"UnmarshalJSON\":{\"In\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"UnmarshalText\":{\"In\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"Value\":{\"In\":[],\"Out\":[{\"Name\":\"Value\",\"Ident\":\"driver.Value\",\"Kind\":20,\"PkgPath\":\"database/sql/driver\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]}}}},\"immutable\":true,\"position\":{\"Index\":0,\"MixedIn\":false,\"MixinIndex\":0},\"schema_type\":{\"mysql\":\"decimal(36,18)\",\"postgres\":\"numeric(36,18)\"},\"annotations\":{\"EntGQL\":{\"Type\":\"String\"}},\"comment\":\"The unit amount of share or money in this account\"},{\"name\":\"value\",\"type\":{\"Type\":20,\"Ident\":\"decimal.Decimal\",\"PkgPath\":\"github.com/shopspring/decimal\",\"PkgName\":\"decimal\",\"Nillable\":false,\"RType\":{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":{\"Abs\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Add\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Atan\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"BigFloat\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"*big.Float\",\"Kind\":22,\"PkgPath\":\"\",\"Methods\":null}]},\"BigInt\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"*big.Int\",\"Kind\":22,\"PkgPath\":\"\",\"Methods\":null}]},\"Ceil\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Cmp\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"int\",\"Ident\":\"int\",\"Kind\":2,\"PkgPath\":\"\",\"Methods\":null}]},\"Coefficient\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"*big.Int\",\"Kind\":22,\"PkgPath\":\"\",\"Methods\":null}]},\"CoefficientInt64\":{\"In\":[],\"Out\":[{\"Name\":\"int64\",\"Ident\":\"int64\",\"Kind\":6,\"PkgPath\":\"\",\"Methods\":null}]},\"Compare\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"int\",\"Ident\":\"int\",\"Kind\":2,\"PkgPath\":\"\",\"Methods\":null}]},\"Copy\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Cos\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Div\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"DivRound\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Equal\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"Equals\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"ExpHullAbrham\":{\"In\":[{\"Name\":\"uint32\",\"Ident\":\"uint32\",\"Kind\":10,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"ExpTaylor\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"Exponent\":{\"In\":[],\"Out\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}]},\"Float64\":{\"In\":[],\"Out\":[{\"Name\":\"float64\",\"Ident\":\"float64\",\"Kind\":14,\"PkgPath\":\"\",\"Methods\":null},{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"Floor\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"GobDecode\":{\"In\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"GobEncode\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"GreaterThan\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"GreaterThanOrEqual\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"InexactFloat64\":{\"In\":[],\"Out\":[{\"Name\":\"float64\",\"Ident\":\"float64\",\"Kind\":14,\"PkgPath\":\"\",\"Methods\":null}]},\"IntPart\":{\"In\":[],\"Out\":[{\"Name\":\"int64\",\"Ident\":\"int64\",\"Kind\":6,\"PkgPath\":\"\",\"Methods\":null}]},\"IsInteger\":{\"In\":[],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"IsNegative\":{\"In\":[],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"IsPositive\":{\"In\":[],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"IsZero\":{\"In\":[],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"LessThan\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"LessThanOrEqual\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"Ln\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"MarshalBinary\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"MarshalJSON\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"MarshalText\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"Mod\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Mul\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Neg\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"NumDigits\":{\"In\":[],\"Out\":[{\"Name\":\"int\",\"Ident\":\"int\",\"Kind\":2,\"PkgPath\":\"\",\"Methods\":null}]},\"Pow\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"PowBigInt\":{\"In\":[{\"Name\":\"\",\"Ident\":\"*big.Int\",\"Kind\":22,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"PowInt32\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"PowWithPrecision\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"QuoRem\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Rat\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"*big.Rat\",\"Kind\":22,\"PkgPath\":\"\",\"Methods\":null}]},\"Round\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"RoundBank\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"RoundCash\":{\"In\":[{\"Name\":\"uint8\",\"Ident\":\"uint8\",\"Kind\":8,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"RoundCeil\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"RoundDown\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"RoundFloor\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"RoundUp\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Scan\":{\"In\":[{\"Name\":\"\",\"Ident\":\"interface {}\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"Shift\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Sign\":{\"In\":[],\"Out\":[{\"Name\":\"int\",\"Ident\":\"int\",\"Kind\":2,\"PkgPath\":\"\",\"Methods\":null}]},\"Sin\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"String\":{\"In\":[],\"Out\":[{\"Name\":\"string\",\"Ident\":\"string\",\"Kind\":24,\"PkgPath\":\"\",\"Methods\":null}]},\"StringFixed\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"string\",\"Ident\":\"string\",\"Kind\":24,\"PkgPath\":\"\",\"Methods\":null}]},\"StringFixedBank\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"string\",\"Ident\":\"string\",\"Kind\":24,\"PkgPath\":\"\",\"Methods\":null}]},\"StringFixedCash\":{\"In\":[{\"Name\":\"uint8\",\"Ident\":\"uint8\",\"Kind\":8,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"string\",\"Ident\":\"string\",\"Kind\":24,\"PkgPath\":\"\",\"Methods\":null}]},\"StringScaled\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"string\",\"Ident\":\"string\",\"Kind\":24,\"PkgPath\":\"\",\"Methods\":null}]},\"Sub\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Tan\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Truncate\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"UnmarshalBinary\":{\"In\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"UnmarshalJSON\":{\"In\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"UnmarshalText\":{\"In\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"Value\":{\"In\":[],\"Out\":[{\"Name\":\"Value\",\"Ident\":\"driver.Value\",\"Kind\":20,\"PkgPath\":\"database/sql/driver\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]}}}},\"immutable\":true,\"position\":{\"Index\":1,\"MixedIn\":false,\"MixinIndex\":0},\"schema_type\":{\"mysql\":\"decimal(18,10)\",\"postgres\":\"numeric(18,10)\"},\"annotations\":{\"EntGQL\":{\"Skip\":48,\"Type\":\"String\"}},\"comment\":\"The value of 1 share in the native currency. If this is just a currency account, then this field will be 1\"},{\"name\":\"balance\",\"type\":{\"Type\":20,\"Ident\":\"decimal.Decimal\",\"PkgPath\":\"github.com/shopspring/decimal\",\"PkgName\":\"decimal\",\"Nillable\":false,\"RType\":{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":{\"Abs\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Add\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Atan\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"BigFloat\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"*big.Float\",\"Kind\":22,\"PkgPath\":\"\",\"Methods\":null}]},\"BigInt\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"*big.Int\",\"Kind\":22,\"PkgPath\":\"\",\"Methods\":null}]},\"Ceil\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Cmp\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"int\",\"Ident\":\"int\",\"Kind\":2,\"PkgPath\":\"\",\"Methods\":null}]},\"Coefficient\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"*big.Int\",\"Kind\":22,\"PkgPath\":\"\",\"Methods\":null}]},\"CoefficientInt64\":{\"In\":[],\"Out\":[{\"Name\":\"int64\",\"Ident\":\"int64\",\"Kind\":6,\"PkgPath\":\"\",\"Methods\":null}]},\"Compare\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"int\",\"Ident\":\"int\",\"Kind\":2,\"PkgPath\":\"\",\"Methods\":null}]},\"Copy\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Cos\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Div\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"DivRound\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Equal\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"Equals\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"ExpHullAbrham\":{\"In\":[{\"Name\":\"uint32\",\"Ident\":\"uint32\",\"Kind\":10,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"ExpTaylor\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"Exponent\":{\"In\":[],\"Out\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}]},\"Float64\":{\"In\":[],\"Out\":[{\"Name\":\"float64\",\"Ident\":\"float64\",\"Kind\":14,\"PkgPath\":\"\",\"Methods\":null},{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"Floor\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"GobDecode\":{\"In\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"GobEncode\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"GreaterThan\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"GreaterThanOrEqual\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"InexactFloat64\":{\"In\":[],\"Out\":[{\"Name\":\"float64\",\"Ident\":\"float64\",\"Kind\":14,\"PkgPath\":\"\",\"Methods\":null}]},\"IntPart\":{\"In\":[],\"Out\":[{\"Name\":\"int64\",\"Ident\":\"int64\",\"Kind\":6,\"PkgPath\":\"\",\"Methods\":null}]},\"IsInteger\":{\"In\":[],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"IsNegative\":{\"In\":[],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"IsPositive\":{\"In\":[],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"IsZero\":{\"In\":[],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"LessThan\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"LessThanOrEqual\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"Ln\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"MarshalBinary\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"MarshalJSON\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"MarshalText\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"Mod\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Mul\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Neg\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"NumDigits\":{\"In\":[],\"Out\":[{\"Name\":\"int\",\"Ident\":\"int\",\"Kind\":2,\"PkgPath\":\"\",\"Methods\":null}]},\"Pow\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"PowBigInt\":{\"In\":[{\"Name\":\"\",\"Ident\":\"*big.Int\",\"Kind\":22,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"PowInt32\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"PowWithPrecision\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"QuoRem\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Rat\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"*big.Rat\",\"Kind\":22,\"PkgPath\":\"\",\"Methods\":null}]},\"Round\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"RoundBank\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"RoundCash\":{\"In\":[{\"Name\":\"uint8\",\"Ident\":\"uint8\",\"Kind\":8,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"RoundCeil\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"RoundDown\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"RoundFloor\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"RoundUp\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Scan\":{\"In\":[{\"Name\":\"\",\"Ident\":\"interface {}\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"Shift\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Sign\":{\"In\":[],\"Out\":[{\"Name\":\"int\",\"Ident\":\"int\",\"Kind\":2,\"PkgPath\":\"\",\"Methods\":null}]},\"Sin\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"String\":{\"In\":[],\"Out\":[{\"Name\":\"string\",\"Ident\":\"string\",\"Kind\":24,\"PkgPath\":\"\",\"Methods\":null}]},\"StringFixed\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"string\",\"Ident\":\"string\",\"Kind\":24,\"PkgPath\":\"\",\"Methods\":null}]},\"StringFixedBank\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"string\",\"Ident\":\"string\",\"Kind\":24,\"PkgPath\":\"\",\"Methods\":null}]},\"StringFixedCash\":{\"In\":[{\"Name\":\"uint8\",\"Ident\":\"uint8\",\"Kind\":8,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"string\",\"Ident\":\"string\",\"Kind\":24,\"PkgPath\":\"\",\"Methods\":null}]},\"StringScaled\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"string\",\"Ident\":\"string\",\"Kind\":24,\"PkgPath\":\"\",\"Methods\":null}]},\"Sub\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Tan\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Truncate\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"UnmarshalBinary\":{\"In\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"UnmarshalJSON\":{\"In\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"UnmarshalText\":{\"In\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"Value\":{\"In\":[],\"Out\":[{\"Name\":\"Value\",\"Ident\":\"driver.Value\",\"Kind\":20,\"PkgPath\":\"database/sql/driver\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]}}}},\"immutable\":true,\"position\":{\"Index\":2,\"MixedIn\":false,\"MixinIndex\":0},\"schema_type\":{\"mysql\":\"decimal(36,18)\",\"postgres\":\"numeric(36,18)\"},\"annotations\":{\"EntGQL\":{\"Skip\":48,\"Type\":\"String\"}},\"comment\":\"The total balance of this account its currency\"}],\"annotations\":{\"EntGQL\":{\"QueryField\":{}},\"EntSQL\":{\"increment_start\":30064771072}}},{\"name\":\"SnapshotFXRate\",\"config\":{\"Table\":\"\"},\"edges\":[{\"name\":\"snapshot\",\"type\":\"Snapshot\",\"ref_name\":\"snapshot_fx_rates\",\"unique\":true,\"inverse\":true,\"required\":true,\"immutable\":true,\"annotations\":{\"EntGQL\":{\"Skip\":48}}}],\"fields\":[{\"name\":\"create_time\",\"type\":{\"Type\":2,\"Ident\":\"\",\"PkgPath\":\"time\",\"PkgName\":\"\",\"Nillable\":false,\"RType\":null},\"default\":true,\"default_kind\":19,\"immutable\":true,\"position\":{\"Index\":0,\"MixedIn\":true,\"MixinIndex\":0}},{\"name\":\"update_time\",\"type\":{\"Type\":2,\"Ident\":\"\",\"PkgPath\":\"time\",\"PkgName\":\"\",\"Nillable\":false,\"RType\":null},\"default\":true,\"default_kind\":19,\"update_default\":true,\"position\":{\"Index\":1,\"MixedIn\":true,\"MixinIndex\":0}},{\"name\":\"from_currency\",\"type\":{\"Type\":7,\"Ident\":\"\",\"PkgPath\":\"\",\"PkgName\":\"\",\"Nillable\":false,\"RType\":null},\"size\":2147483647,\"immutable\":true,\"position\":{\"Index\":0,\"MixedIn\":false,\"MixinIndex\":0}},{\"name\":\"to_currency\",\"type\":{\"Type\":7,\"Ident\":\"\",\"PkgPath\":\"\",\"PkgName\":\"\",\"Nillable\":false,\"RType\":null},\"size\":2147483647,\"immutable\":true,\"position\":{\"Index\":1,\"MixedIn\":false,\"MixinIndex\":0}},{\"name\":\"fx_rate\",\"type\":{\"Type\":20,\"Ident\":\"decimal.Decimal\",\"PkgPath\":\"github.com/shopspring/decimal\",\"PkgName\":\"decimal\",\"Nillable\":false,\"RType\":{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":{\"Abs\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Add\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Atan\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"BigFloat\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"*big.Float\",\"Kind\":22,\"PkgPath\":\"\",\"Methods\":null}]},\"BigInt\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"*big.Int\",\"Kind\":22,\"PkgPath\":\"\",\"Methods\":null}]},\"Ceil\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Cmp\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"int\",\"Ident\":\"int\",\"Kind\":2,\"PkgPath\":\"\",\"Methods\":null}]},\"Coefficient\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"*big.Int\",\"Kind\":22,\"PkgPath\":\"\",\"Methods\":null}]},\"CoefficientInt64\":{\"In\":[],\"Out\":[{\"Name\":\"int64\",\"Ident\":\"int64\",\"Kind\":6,\"PkgPath\":\"\",\"Methods\":null}]},\"Compare\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"int\",\"Ident\":\"int\",\"Kind\":2,\"PkgPath\":\"\",\"Methods\":null}]},\"Copy\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Cos\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Div\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"DivRound\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Equal\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"Equals\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"ExpHullAbrham\":{\"In\":[{\"Name\":\"uint32\",\"Ident\":\"uint32\",\"Kind\":10,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"ExpTaylor\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"Exponent\":{\"In\":[],\"Out\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}]},\"Float64\":{\"In\":[],\"Out\":[{\"Name\":\"float64\",\"Ident\":\"float64\",\"Kind\":14,\"PkgPath\":\"\",\"Methods\":null},{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"Floor\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"GobDecode\":{\"In\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"GobEncode\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"GreaterThan\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"GreaterThanOrEqual\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"InexactFloat64\":{\"In\":[],\"Out\":[{\"Name\":\"float64\",\"Ident\":\"float64\",\"Kind\":14,\"PkgPath\":\"\",\"Methods\":null}]},\"IntPart\":{\"In\":[],\"Out\":[{\"Name\":\"int64\",\"Ident\":\"int64\",\"Kind\":6,\"PkgPath\":\"\",\"Methods\":null}]},\"IsInteger\":{\"In\":[],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"IsNegative\":{\"In\":[],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"IsPositive\":{\"In\":[],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"IsZero\":{\"In\":[],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"LessThan\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"LessThanOrEqual\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"Ln\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"MarshalBinary\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"MarshalJSON\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"MarshalText\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"Mod\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Mul\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Neg\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"NumDigits\":{\"In\":[],\"Out\":[{\"Name\":\"int\",\"Ident\":\"int\",\"Kind\":2,\"PkgPath\":\"\",\"Methods\":null}]},\"Pow\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"PowBigInt\":{\"In\":[{\"Name\":\"\",\"Ident\":\"*big.Int\",\"Kind\":22,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"PowInt32\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"PowWithPrecision\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"QuoRem\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Rat\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"*big.Rat\",\"Kind\":22,\"PkgPath\":\"\",\"Methods\":null}]},\"Round\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"RoundBank\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"RoundCash\":{\"In\":[{\"Name\":\"uint8\",\"Ident\":\"uint8\",\"Kind\":8,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"RoundCeil\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"RoundDown\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"RoundFloor\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"RoundUp\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Scan\":{\"In\":[{\"Name\":\"\",\"Ident\":\"interface {}\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"Shift\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Sign\":{\"In\":[],\"Out\":[{\"Name\":\"int\",\"Ident\":\"int\",\"Kind\":2,\"PkgPath\":\"\",\"Methods\":null}]},\"Sin\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"String\":{\"In\":[],\"Out\":[{\"Name\":\"string\",\"Ident\":\"string\",\"Kind\":24,\"PkgPath\":\"\",\"Methods\":null}]},\"StringFixed\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"string\",\"Ident\":\"string\",\"Kind\":24,\"PkgPath\":\"\",\"Methods\":null}]},\"StringFixedBank\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"string\",\"Ident\":\"string\",\"Kind\":24,\"PkgPath\":\"\",\"Methods\":null}]},\"StringFixedCash\":{\"In\":[{\"Name\":\"uint8\",\"Ident\":\"uint8\",\"Kind\":8,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"string\",\"Ident\":\"string\",\"Kind\":24,\"PkgPath\":\"\",\"Methods\":null}]},\"StringScaled\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"string\",\"Ident\":\"string\",\"Kind\":24,\"PkgPath\":\"\",\"Methods\":null}]},\"Sub\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Tan\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Truncate\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"UnmarshalBinary\":{\"In\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"UnmarshalJSON\":{\"In\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"UnmarshalText\":{\"In\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"Value\":{\"In\":[],\"Out\":[{\"Name\":\"Value\",\"Ident\":\"driver.Value\",\"Kind\":20,\"PkgPath\":\"database/sql/driver\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]}}}},\"immutable\":true,\"position\":{\"Index\":2,\"MixedIn\":false,\"MixinIndex\":0},\"schema_type\":{\"mysql\":\"decimal(18,10)\",\"postgres\":\"numeric(18,10)\"},\"annotations\":{\"EntGQL\":{\"Skip\":48,\"Type\":\"String\"}},\"comment\":\"The exchange rate from the native currency to user's default display currency\"}],\"annotations\":{\"EntGQL\":{\"QueryField\":{}},\"EntSQL\":{\"increment_start\":34359738368}}},{\"name\":\"Transaction\",\"config\":{\"Table\":\"\"},\"edges\":[{\"name\":\"profile\",\"type\":\"Profile\",\"ref_name\":\"transactions\",\"unique\":true,\"inverse\":true,\"required\":true,\"annotations\":{\"EntGQL\":{\"Skip\":48}}},{\"name\":\"category\",\"type\":\"Category\",\"ref_name\":\"transactions\",\"unique\":true,\"inverse\":true,\"required\":true,\"annotations\":{\"EntGQL\":{\"Skip\":48}}},{\"name\":\"transaction_entries\",\"type\":\"TransactionEntry\"}],\"fields\":[{\"name\":\"create_time\",\"type\":{\"Type\":2,\"Ident\":\"\",\"PkgPath\":\"time\",\"PkgName\":\"\",\"Nillable\":false,\"RType\":null},\"default\":true,\"default_kind\":19,\"immutable\":true,\"position\":{\"Index\":0,\"MixedIn\":true,\"MixinIndex\":0}},{\"name\":\"update_time\",\"type\":{\"Type\":2,\"Ident\":\"\",\"PkgPath\":\"time\",\"PkgName\":\"\",\"Nillable\":false,\"RType\":null},\"default\":true,\"default_kind\":19,\"update_default\":true,\"position\":{\"Index\":1,\"MixedIn\":true,\"MixinIndex\":0}},{\"name\":\"note\",\"type\":{\"Type\":7,\"Ident\":\"\",\"PkgPath\":\"\",\"PkgName\":\"\",\"Nillable\":false,\"RType\":null},\"size\":2147483647,\"optional\":true,\"position\":{\"Index\":0,\"MixedIn\":false,\"MixinIndex\":0}}],\"annotations\":{\"EntGQL\":{\"MutationInputs\":[{\"IsCreate\":true},{}],\"QueryField\":{},\"RelayConnection\":true},\"EntSQL\":{\"increment_start\":8589934592}}},{\"name\":\"TransactionEntry\",\"config\":{\"Table\":\"\"},\"edges\":[{\"name\":\"account\",\"type\":\"Account\",\"ref_name\":\"transaction_entries\",\"unique\":true,\"inverse\":true,\"required\":true},{\"name\":\"transaction\",\"type\":\"Transaction\",\"ref_name\":\"transaction_entries\",\"unique\":true,\"inverse\":true,\"required\":true}],\"fields\":[{\"name\":\"create_time\",\"type\":{\"Type\":2,\"Ident\":\"\",\"PkgPath\":\"time\",\"PkgName\":\"\",\"Nillable\":false,\"RType\":null},\"default\":true,\"default_kind\":19,\"immutable\":true,\"position\":{\"Index\":0,\"MixedIn\":true,\"MixinIndex\":0}},{\"name\":\"update_time\",\"type\":{\"Type\":2,\"Ident\":\"\",\"PkgPath\":\"time\",\"PkgName\":\"\",\"Nillable\":false,\"RType\":null},\"default\":true,\"default_kind\":19,\"update_default\":true,\"position\":{\"Index\":1,\"MixedIn\":true,\"MixinIndex\":0}},{\"name\":\"note\",\"type\":{\"Type\":7,\"Ident\":\"\",\"PkgPath\":\"\",\"PkgName\":\"\",\"Nillable\":false,\"RType\":null},\"size\":2147483647,\"optional\":true,\"position\":{\"Index\":0,\"MixedIn\":false,\"MixinIndex\":0}},{\"name\":\"amount\",\"type\":{\"Type\":20,\"Ident\":\"decimal.Decimal\",\"PkgPath\":\"github.com/shopspring/decimal\",\"PkgName\":\"decimal\",\"Nillable\":false,\"RType\":{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":{\"Abs\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Add\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Atan\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"BigFloat\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"*big.Float\",\"Kind\":22,\"PkgPath\":\"\",\"Methods\":null}]},\"BigInt\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"*big.Int\",\"Kind\":22,\"PkgPath\":\"\",\"Methods\":null}]},\"Ceil\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Cmp\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"int\",\"Ident\":\"int\",\"Kind\":2,\"PkgPath\":\"\",\"Methods\":null}]},\"Coefficient\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"*big.Int\",\"Kind\":22,\"PkgPath\":\"\",\"Methods\":null}]},\"CoefficientInt64\":{\"In\":[],\"Out\":[{\"Name\":\"int64\",\"Ident\":\"int64\",\"Kind\":6,\"PkgPath\":\"\",\"Methods\":null}]},\"Compare\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"int\",\"Ident\":\"int\",\"Kind\":2,\"PkgPath\":\"\",\"Methods\":null}]},\"Copy\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Cos\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Div\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"DivRound\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Equal\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"Equals\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"ExpHullAbrham\":{\"In\":[{\"Name\":\"uint32\",\"Ident\":\"uint32\",\"Kind\":10,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"ExpTaylor\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"Exponent\":{\"In\":[],\"Out\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}]},\"Float64\":{\"In\":[],\"Out\":[{\"Name\":\"float64\",\"Ident\":\"float64\",\"Kind\":14,\"PkgPath\":\"\",\"Methods\":null},{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"Floor\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"GobDecode\":{\"In\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"GobEncode\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"GreaterThan\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"GreaterThanOrEqual\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"InexactFloat64\":{\"In\":[],\"Out\":[{\"Name\":\"float64\",\"Ident\":\"float64\",\"Kind\":14,\"PkgPath\":\"\",\"Methods\":null}]},\"IntPart\":{\"In\":[],\"Out\":[{\"Name\":\"int64\",\"Ident\":\"int64\",\"Kind\":6,\"PkgPath\":\"\",\"Methods\":null}]},\"IsInteger\":{\"In\":[],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"IsNegative\":{\"In\":[],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"IsPositive\":{\"In\":[],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"IsZero\":{\"In\":[],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"LessThan\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"LessThanOrEqual\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"Ln\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"MarshalBinary\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"MarshalJSON\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"MarshalText\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"Mod\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Mul\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Neg\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"NumDigits\":{\"In\":[],\"Out\":[{\"Name\":\"int\",\"Ident\":\"int\",\"Kind\":2,\"PkgPath\":\"\",\"Methods\":null}]},\"Pow\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"PowBigInt\":{\"In\":[{\"Name\":\"\",\"Ident\":\"*big.Int\",\"Kind\":22,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"PowInt32\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"PowWithPrecision\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"QuoRem\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Rat\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"*big.Rat\",\"Kind\":22,\"PkgPath\":\"\",\"Methods\":null}]},\"Round\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"RoundBank\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"RoundCash\":{\"In\":[{\"Name\":\"uint8\",\"Ident\":\"uint8\",\"Kind\":8,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"RoundCeil\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"RoundDown\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"RoundFloor\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"RoundUp\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Scan\":{\"In\":[{\"Name\":\"\",\"Ident\":\"interface {}\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"Shift\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Sign\":{\"In\":[],\"Out\":[{\"Name\":\"int\",\"Ident\":\"int\",\"Kind\":2,\"PkgPath\":\"\",\"Methods\":null}]},\"Sin\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"String\":{\"In\":[],\"Out\":[{\"Name\":\"string\",\"Ident\":\"string\",\"Kind\":24,\"PkgPath\":\"\",\"Methods\":null}]},\"StringFixed\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"string\",\"Ident\":\"string\",\"Kind\":24,\"PkgPath\":\"\",\"Methods\":null}]},\"StringFixedBank\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"string\",\"Ident\":\"string\",\"Kind\":24,\"PkgPath\":\"\",\"Methods\":null}]},\"StringFixedCash\":{\"In\":[{\"Name\":\"uint8\",\"Ident\":\"uint8\",\"Kind\":8,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"string\",\"Ident\":\"string\",\"Kind\":24,\"PkgPath\":\"\",\"Methods\":null}]},\"StringScaled\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"string\",\"Ident\":\"string\",\"Kind\":24,\"PkgPath\":\"\",\"Methods\":null}]},\"Sub\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Tan\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Truncate\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"UnmarshalBinary\":{\"In\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"UnmarshalJSON\":{\"In\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"UnmarshalText\":{\"In\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"Value\":{\"In\":[],\"Out\":[{\"Name\":\"Value\",\"Ident\":\"driver.Value\",\"Kind\":20,\"PkgPath\":\"database/sql/driver\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]}}}},\"position\":{\"Index\":1,\"MixedIn\":false,\"MixinIndex\":0},\"schema_type\":{\"mysql\":\"decimal(36,18)\",\"postgres\":\"numeric(36,18)\"},\"annotations\":{\"EntGQL\":{\"Type\":\"String\"}},\"comment\":\"The unit amount of share or money in this transaction entry\"},{\"name\":\"value\",\"type\":{\"Type\":20,\"Ident\":\"decimal.Decimal\",\"PkgPath\":\"github.com/shopspring/decimal\",\"PkgName\":\"decimal\",\"Nillable\":false,\"RType\":{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":{\"Abs\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Add\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Atan\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"BigFloat\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"*big.Float\",\"Kind\":22,\"PkgPath\":\"\",\"Methods\":null}]},\"BigInt\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"*big.Int\",\"Kind\":22,\"PkgPath\":\"\",\"Methods\":null}]},\"Ceil\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Cmp\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"int\",\"Ident\":\"int\",\"Kind\":2,\"PkgPath\":\"\",\"Methods\":null}]},\"Coefficient\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"*big.Int\",\"Kind\":22,\"PkgPath\":\"\",\"Methods\":null}]},\"CoefficientInt64\":{\"In\":[],\"Out\":[{\"Name\":\"int64\",\"Ident\":\"int64\",\"Kind\":6,\"PkgPath\":\"\",\"Methods\":null}]},\"Compare\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"int\",\"Ident\":\"int\",\"Kind\":2,\"PkgPath\":\"\",\"Methods\":null}]},\"Copy\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Cos\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Div\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"DivRound\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Equal\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"Equals\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"ExpHullAbrham\":{\"In\":[{\"Name\":\"uint32\",\"Ident\":\"uint32\",\"Kind\":10,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"ExpTaylor\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"Exponent\":{\"In\":[],\"Out\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}]},\"Float64\":{\"In\":[],\"Out\":[{\"Name\":\"float64\",\"Ident\":\"float64\",\"Kind\":14,\"PkgPath\":\"\",\"Methods\":null},{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"Floor\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"GobDecode\":{\"In\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"GobEncode\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"GreaterThan\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"GreaterThanOrEqual\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"InexactFloat64\":{\"In\":[],\"Out\":[{\"Name\":\"float64\",\"Ident\":\"float64\",\"Kind\":14,\"PkgPath\":\"\",\"Methods\":null}]},\"IntPart\":{\"In\":[],\"Out\":[{\"Name\":\"int64\",\"Ident\":\"int64\",\"Kind\":6,\"PkgPath\":\"\",\"Methods\":null}]},\"IsInteger\":{\"In\":[],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"IsNegative\":{\"In\":[],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"IsPositive\":{\"In\":[],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"IsZero\":{\"In\":[],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"LessThan\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"LessThanOrEqual\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"Ln\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"MarshalBinary\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"MarshalJSON\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"MarshalText\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"Mod\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Mul\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Neg\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"NumDigits\":{\"In\":[],\"Out\":[{\"Name\":\"int\",\"Ident\":\"int\",\"Kind\":2,\"PkgPath\":\"\",\"Methods\":null}]},\"Pow\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"PowBigInt\":{\"In\":[{\"Name\":\"\",\"Ident\":\"*big.Int\",\"Kind\":22,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"PowInt32\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"PowWithPrecision\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"QuoRem\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Rat\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"*big.Rat\",\"Kind\":22,\"PkgPath\":\"\",\"Methods\":null}]},\"Round\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"RoundBank\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"RoundCash\":{\"In\":[{\"Name\":\"uint8\",\"Ident\":\"uint8\",\"Kind\":8,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"RoundCeil\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"RoundDown\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"RoundFloor\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"RoundUp\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Scan\":{\"In\":[{\"Name\":\"\",\"Ident\":\"interface {}\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"Shift\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Sign\":{\"In\":[],\"Out\":[{\"Name\":\"int\",\"Ident\":\"int\",\"Kind\":2,\"PkgPath\":\"\",\"Methods\":null}]},\"Sin\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"String\":{\"In\":[],\"Out\":[{\"Name\":\"string\",\"Ident\":\"string\",\"Kind\":24,\"PkgPath\":\"\",\"Methods\":null}]},\"StringFixed\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"string\",\"Ident\":\"string\",\"Kind\":24,\"PkgPath\":\"\",\"Methods\":null}]},\"StringFixedBank\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"string\",\"Ident\":\"string\",\"Kind\":24,\"PkgPath\":\"\",\"Methods\":null}]},\"StringFixedCash\":{\"In\":[{\"Name\":\"uint8\",\"Ident\":\"uint8\",\"Kind\":8,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"string\",\"Ident\":\"string\",\"Kind\":24,\"PkgPath\":\"\",\"Methods\":null}]},\"StringScaled\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"string\",\"Ident\":\"string\",\"Kind\":24,\"PkgPath\":\"\",\"Methods\":null}]},\"Sub\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Tan\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Truncate\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"UnmarshalBinary\":{\"In\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"UnmarshalJSON\":{\"In\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"UnmarshalText\":{\"In\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"Value\":{\"In\":[],\"Out\":[{\"Name\":\"Value\",\"Ident\":\"driver.Value\",\"Kind\":20,\"PkgPath\":\"database/sql/driver\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]}}}},\"position\":{\"Index\":2,\"MixedIn\":false,\"MixinIndex\":0},\"schema_type\":{\"mysql\":\"decimal(18,10)\",\"postgres\":\"numeric(18,10)\"},\"annotations\":{\"EntGQL\":{\"Skip\":48,\"Type\":\"String\"}},\"comment\":\"The value of 1 share in the native currency. If this is just a currency account, then this field will be 1\"},{\"name\":\"balance\",\"type\":{\"Type\":20,\"Ident\":\"decimal.Decimal\",\"PkgPath\":\"github.com/shopspring/decimal\",\"PkgName\":\"decimal\",\"Nillable\":false,\"RType\":{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":{\"Abs\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Add\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Atan\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"BigFloat\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"*big.Float\",\"Kind\":22,\"PkgPath\":\"\",\"Methods\":null}]},\"BigInt\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"*big.Int\",\"Kind\":22,\"PkgPath\":\"\",\"Methods\":null}]},\"Ceil\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Cmp\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"int\",\"Ident\":\"int\",\"Kind\":2,\"PkgPath\":\"\",\"Methods\":null}]},\"Coefficient\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"*big.Int\",\"Kind\":22,\"PkgPath\":\"\",\"Methods\":null}]},\"CoefficientInt64\":{\"In\":[],\"Out\":[{\"Name\":\"int64\",\"Ident\":\"int64\",\"Kind\":6,\"PkgPath\":\"\",\"Methods\":null}]},\"Compare\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"int\",\"Ident\":\"int\",\"Kind\":2,\"PkgPath\":\"\",\"Methods\":null}]},\"Copy\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Cos\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Div\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"DivRound\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Equal\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"Equals\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"ExpHullAbrham\":{\"In\":[{\"Name\":\"uint32\",\"Ident\":\"uint32\",\"Kind\":10,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"ExpTaylor\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"Exponent\":{\"In\":[],\"Out\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}]},\"Float64\":{\"In\":[],\"Out\":[{\"Name\":\"float64\",\"Ident\":\"float64\",\"Kind\":14,\"PkgPath\":\"\",\"Methods\":null},{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"Floor\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"GobDecode\":{\"In\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"GobEncode\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"GreaterThan\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"GreaterThanOrEqual\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"InexactFloat64\":{\"In\":[],\"Out\":[{\"Name\":\"float64\",\"Ident\":\"float64\",\"Kind\":14,\"PkgPath\":\"\",\"Methods\":null}]},\"IntPart\":{\"In\":[],\"Out\":[{\"Name\":\"int64\",\"Ident\":\"int64\",\"Kind\":6,\"PkgPath\":\"\",\"Methods\":null}]},\"IsInteger\":{\"In\":[],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"IsNegative\":{\"In\":[],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"IsPositive\":{\"In\":[],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"IsZero\":{\"In\":[],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"LessThan\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"LessThanOrEqual\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"Ln\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"MarshalBinary\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"MarshalJSON\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"MarshalText\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"Mod\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Mul\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Neg\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"NumDigits\":{\"In\":[],\"Out\":[{\"Name\":\"int\",\"Ident\":\"int\",\"Kind\":2,\"PkgPath\":\"\",\"Methods\":null}]},\"Pow\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"PowBigInt\":{\"In\":[{\"Name\":\"\",\"Ident\":\"*big.Int\",\"Kind\":22,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"PowInt32\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"PowWithPrecision\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"QuoRem\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Rat\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"*big.Rat\",\"Kind\":22,\"PkgPath\":\"\",\"Methods\":null}]},\"Round\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"RoundBank\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"RoundCash\":{\"In\":[{\"Name\":\"uint8\",\"Ident\":\"uint8\",\"Kind\":8,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"RoundCeil\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"RoundDown\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"RoundFloor\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"RoundUp\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Scan\":{\"In\":[{\"Name\":\"\",\"Ident\":\"interface {}\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"Shift\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Sign\":{\"In\":[],\"Out\":[{\"Name\":\"int\",\"Ident\":\"int\",\"Kind\":2,\"PkgPath\":\"\",\"Methods\":null}]},\"Sin\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"String\":{\"In\":[],\"Out\":[{\"Name\":\"string\",\"Ident\":\"string\",\"Kind\":24,\"PkgPath\":\"\",\"Methods\":null}]},\"StringFixed\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"string\",\"Ident\":\"string\",\"Kind\":24,\"PkgPath\":\"\",\"Methods\":null}]},\"StringFixedBank\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"string\",\"Ident\":\"string\",\"Kind\":24,\"PkgPath\":\"\",\"Methods\":null}]},\"StringFixedCash\":{\"In\":[{\"Name\":\"uint8\",\"Ident\":\"uint8\",\"Kind\":8,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"string\",\"Ident\":\"string\",\"Kind\":24,\"PkgPath\":\"\",\"Methods\":null}]},\"StringScaled\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"string\",\"Ident\":\"string\",\"Kind\":24,\"PkgPath\":\"\",\"Methods\":null}]},\"Sub\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Tan\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Truncate\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"UnmarshalBinary\":{\"In\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"UnmarshalJSON\":{\"In\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"UnmarshalText\":{\"In\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"Value\":{\"In\":[],\"Out\":[{\"Name\":\"Value\",\"Ident\":\"driver.Value\",\"Kind\":20,\"PkgPath\":\"database/sql/driver\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]}}}},\"position\":{\"Index\":3,\"MixedIn\":false,\"MixinIndex\":0},\"schema_type\":{\"mysql\":\"decimal(36,18)\",\"postgres\":\"numeric(36,18)\"},\"annotations\":{\"EntGQL\":{\"Skip\":48,\"Type\":\"String\"}},\"comment\":\"The total balance of this transaction entry in this account's currency\"}],\"annotations\":{\"EntGQL\":{\"MutationInputs\":[{\"IsCreate\":true},{}]},\"EntSQL\":{\"increment_start\":12884901888}}},{\"name\":\"User\",\"config\":{\"Table\":\"\"},\"edges\":[{\"name\":\"user_keys\",\"type\":\"UserKey\",\"annotations\":{\"EntSQL\":{\"on_delete\":\"CASCADE\"}}},{\"name\":\"profiles\",\"type\":\"Profile\",\"through\":{\"N\":\"user_profiles\",\"T\":\"UserProfile\"}}],\"fields\":[{\"name\":\"create_time\",\"type\":{\"Type\":2,\"Ident\":\"\",\"PkgPath\":\"time\",\"PkgName\":\"\",\"Nillable\":false,\"RType\":null},\"default\":true,\"default_kind\":19,\"immutable\":true,\"position\":{\"Index\":0,\"MixedIn\":true,\"MixinIndex\":0}},{\"name\":\"update_time\",\"type\":{\"Type\":2,\"Ident\":\"\",\"PkgPath\":\"time\",\"PkgName\":\"\",\"Nillable\":false,\"RType\":null},\"default\":true,\"default_kind\":19,\"update_default\":true,\"position\":{\"Index\":1,\"MixedIn\":true,\"MixinIndex\":0}},{\"name\":\"email\",\"type\":{\"Type\":7,\"Ident\":\"\",\"PkgPath\":\"\",\"PkgName\":\"\",\"Nillable\":false,\"RType\":null},\"unique\":true,\"validators\":1,\"position\":{\"Index\":0,\"MixedIn\":false,\"MixinIndex\":0}}],\"annotations\":{\"EntGQL\":{\"MutationInputs\":[{\"IsCreate\":true}]},\"EntSQL\":{\"increment_start\":17179869184}}},{\"name\":\"UserKey\",\"config\":{\"Table\":\"\"},\"edges\":[{\"name\":\"user\",\"type\":\"User\",\"ref_name\":\"user_keys\",\"unique\":true,\"inverse\":true,\"required\":true}],\"fields\":[{\"name\":\"create_time\",\"type\":{\"Type\":2,\"Ident\":\"\",\"PkgPath\":\"time\",\"PkgName\":\"\",\"Nillable\":false,\"RType\":null},\"default\":true,\"default_kind\":19,\"immutable\":true,\"position\":{\"Index\":0,\"MixedIn\":true,\"MixinIndex\":0}},{\"name\":\"update_time\",\"type\":{\"Type\":2,\"Ident\":\"\",\"PkgPath\":\"time\",\"PkgName\":\"\",\"Nillable\":false,\"RType\":null},\"default\":true,\"default_kind\":19,\"update_default\":true,\"position\":{\"Index\":1,\"MixedIn\":true,\"MixinIndex\":0}},{\"name\":\"key\",\"type\":{\"Type\":7,\"Ident\":\"\",\"PkgPath\":\"\",\"PkgName\":\"\",\"Nillable\":false,\"RType\":null},\"validators\":1,\"position\":{\"Index\":0,\"MixedIn\":false,\"MixinIndex\":0}},{\"name\":\"hashed_password\",\"type\":{\"Type\":7,\"Ident\":\"\",\"PkgPath\":\"\",\"PkgName\":\"\",\"Nillable\":false,\"RType\":null},\"optional\":true,\"position\":{\"Index\":1,\"MixedIn\":false,\"MixinIndex\":0}}],\"annotations\":{\"EntGQL\":{\"MutationInputs\":[{\"IsCreate\":true}]},\"EntSQL\":{\"increment_start\":21474836480}}},{\"name\":\"UserProfile\",\"config\":{\"Table\":\"\"},\"edges\":[{\"name\":\"user\",\"type\":\"User\",\"field\":\"user_id\",\"unique\":true,\"required\":true},{\"name\":\"profile\",\"type\":\"Profile\",\"field\":\"profile_id\",\"unique\":true,\"required\":true}],\"fields\":[{\"name\":\"permission\",\"type\":{\"Type\":6,\"Ident\":\"userprofile.Permission\",\"PkgPath\":\"\",\"PkgName\":\"\",\"Nillable\":false,\"RType\":null},\"enums\":[{\"N\":\"owner\",\"V\":\"owner\"},{\"N\":\"admin\",\"V\":\"admin\"},{\"N\":\"viewer\",\"V\":\"viewer\"}],\"position\":{\"Index\":0,\"MixedIn\":false,\"MixinIndex\":0}},{\"name\":\"user_id\",\"type\":{\"Type\":12,\"Ident\":\"\",\"PkgPath\":\"\",\"PkgName\":\"\",\"Nillable\":false,\"RType\":null},\"position\":{\"Index\":1,\"MixedIn\":false,\"MixinIndex\":0}},{\"name\":\"profile_id\",\"type\":{\"Type\":12,\"Ident\":\"\",\"PkgPath\":\"\",\"PkgName\":\"\",\"Nillable\":false,\"RType\":null},\"position\":{\"Index\":2,\"MixedIn\":false,\"MixinIndex\":0}}],\"annotations\":{\"EntSQL\":{\"increment_start\":42949672960},\"Fields\":{\"ID\":[\"user_id\",\"profile_id\"],\"StructTag\":null}}}],\"Features\":[\"sql/globalid\",\"privacy\",\"schema/snapshot\",\"namedges\"]}"
+const Schema = "{\"Schema\":\"fijoy/ent/schema\",\"Package\":\"fijoy/ent\",\"Schemas\":[{\"name\":\"Account\",\"config\":{\"Table\":\"\"},\"edges\":[{\"name\":\"profile\",\"type\":\"Profile\",\"ref_name\":\"accounts\",\"unique\":true,\"inverse\":true,\"required\":true,\"annotations\":{\"EntGQL\":{\"Skip\":48}}},{\"name\":\"transaction_entries\",\"type\":\"TransactionEntry\"},{\"name\":\"snapshot_accounts\",\"type\":\"SnapshotAccount\"}],\"fields\":[{\"name\":\"create_time\",\"type\":{\"Type\":2,\"Ident\":\"\",\"PkgPath\":\"time\",\"PkgName\":\"\",\"Nillable\":false,\"RType\":null},\"default\":true,\"default_kind\":19,\"immutable\":true,\"position\":{\"Index\":0,\"MixedIn\":true,\"MixinIndex\":0}},{\"name\":\"update_time\",\"type\":{\"Type\":2,\"Ident\":\"\",\"PkgPath\":\"time\",\"PkgName\":\"\",\"Nillable\":false,\"RType\":null},\"default\":true,\"default_kind\":19,\"update_default\":true,\"position\":{\"Index\":1,\"MixedIn\":true,\"MixinIndex\":0}},{\"name\":\"name\",\"type\":{\"Type\":7,\"Ident\":\"\",\"PkgPath\":\"\",\"PkgName\":\"\",\"Nillable\":false,\"RType\":null},\"validators\":1,\"position\":{\"Index\":0,\"MixedIn\":false,\"MixinIndex\":0}},{\"name\":\"institution\",\"type\":{\"Type\":7,\"Ident\":\"\",\"PkgPath\":\"\",\"PkgName\":\"\",\"Nillable\":false,\"RType\":null},\"validators\":1,\"position\":{\"Index\":1,\"MixedIn\":false,\"MixinIndex\":0},\"comment\":\"The financial institution that holds this account\"},{\"name\":\"account_type\",\"type\":{\"Type\":6,\"Ident\":\"account.AccountType\",\"PkgPath\":\"\",\"PkgName\":\"\",\"Nillable\":false,\"RType\":null},\"enums\":[{\"N\":\"liquidity\",\"V\":\"liquidity\"},{\"N\":\"investment\",\"V\":\"investment\"},{\"N\":\"property\",\"V\":\"property\"},{\"N\":\"receivable\",\"V\":\"receivable\"},{\"N\":\"liability\",\"V\":\"liability\"}],\"immutable\":true,\"position\":{\"Index\":2,\"MixedIn\":false,\"MixinIndex\":0}},{\"name\":\"investment_type\",\"type\":{\"Type\":6,\"Ident\":\"account.InvestmentType\",\"PkgPath\":\"\",\"PkgName\":\"\",\"Nillable\":false,\"RType\":null},\"enums\":[{\"N\":\"non_investment\",\"V\":\"non_investment\"},{\"N\":\"taxable\",\"V\":\"taxable\"},{\"N\":\"rrsp\",\"V\":\"rrsp\"},{\"N\":\"tfsa\",\"V\":\"tfsa\"},{\"N\":\"fhsa\",\"V\":\"fhsa\"}],\"position\":{\"Index\":3,\"MixedIn\":false,\"MixinIndex\":0}},{\"name\":\"currency_code\",\"type\":{\"Type\":7,\"Ident\":\"\",\"PkgPath\":\"\",\"PkgName\":\"\",\"Nillable\":false,\"RType\":null},\"immutable\":true,\"validators\":1,\"position\":{\"Index\":4,\"MixedIn\":false,\"MixinIndex\":0}},{\"name\":\"ticker\",\"type\":{\"Type\":7,\"Ident\":\"\",\"PkgPath\":\"\",\"PkgName\":\"\",\"Nillable\":false,\"RType\":null},\"immutable\":true,\"validators\":1,\"position\":{\"Index\":5,\"MixedIn\":false,\"MixinIndex\":0}},{\"name\":\"ticker_type\",\"type\":{\"Type\":6,\"Ident\":\"account.TickerType\",\"PkgPath\":\"\",\"PkgName\":\"\",\"Nillable\":false,\"RType\":null},\"enums\":[{\"N\":\"currency\",\"V\":\"currency\"},{\"N\":\"stock\",\"V\":\"stock\"},{\"N\":\"crypto\",\"V\":\"crypto\"}],\"immutable\":true,\"position\":{\"Index\":6,\"MixedIn\":false,\"MixinIndex\":0}},{\"name\":\"amount\",\"type\":{\"Type\":20,\"Ident\":\"decimal.Decimal\",\"PkgPath\":\"github.com/shopspring/decimal\",\"PkgName\":\"decimal\",\"Nillable\":false,\"RType\":{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":{\"Abs\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Add\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Atan\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"BigFloat\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"*big.Float\",\"Kind\":22,\"PkgPath\":\"\",\"Methods\":null}]},\"BigInt\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"*big.Int\",\"Kind\":22,\"PkgPath\":\"\",\"Methods\":null}]},\"Ceil\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Cmp\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"int\",\"Ident\":\"int\",\"Kind\":2,\"PkgPath\":\"\",\"Methods\":null}]},\"Coefficient\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"*big.Int\",\"Kind\":22,\"PkgPath\":\"\",\"Methods\":null}]},\"CoefficientInt64\":{\"In\":[],\"Out\":[{\"Name\":\"int64\",\"Ident\":\"int64\",\"Kind\":6,\"PkgPath\":\"\",\"Methods\":null}]},\"Compare\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"int\",\"Ident\":\"int\",\"Kind\":2,\"PkgPath\":\"\",\"Methods\":null}]},\"Copy\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Cos\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Div\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"DivRound\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Equal\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"Equals\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"ExpHullAbrham\":{\"In\":[{\"Name\":\"uint32\",\"Ident\":\"uint32\",\"Kind\":10,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"ExpTaylor\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"Exponent\":{\"In\":[],\"Out\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}]},\"Float64\":{\"In\":[],\"Out\":[{\"Name\":\"float64\",\"Ident\":\"float64\",\"Kind\":14,\"PkgPath\":\"\",\"Methods\":null},{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"Floor\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"GobDecode\":{\"In\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"GobEncode\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"GreaterThan\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"GreaterThanOrEqual\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"InexactFloat64\":{\"In\":[],\"Out\":[{\"Name\":\"float64\",\"Ident\":\"float64\",\"Kind\":14,\"PkgPath\":\"\",\"Methods\":null}]},\"IntPart\":{\"In\":[],\"Out\":[{\"Name\":\"int64\",\"Ident\":\"int64\",\"Kind\":6,\"PkgPath\":\"\",\"Methods\":null}]},\"IsInteger\":{\"In\":[],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"IsNegative\":{\"In\":[],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"IsPositive\":{\"In\":[],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"IsZero\":{\"In\":[],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"LessThan\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"LessThanOrEqual\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"Ln\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"MarshalBinary\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"MarshalJSON\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"MarshalText\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"Mod\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Mul\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Neg\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"NumDigits\":{\"In\":[],\"Out\":[{\"Name\":\"int\",\"Ident\":\"int\",\"Kind\":2,\"PkgPath\":\"\",\"Methods\":null}]},\"Pow\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"PowBigInt\":{\"In\":[{\"Name\":\"\",\"Ident\":\"*big.Int\",\"Kind\":22,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"PowInt32\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"PowWithPrecision\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"QuoRem\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Rat\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"*big.Rat\",\"Kind\":22,\"PkgPath\":\"\",\"Methods\":null}]},\"Round\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"RoundBank\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"RoundCash\":{\"In\":[{\"Name\":\"uint8\",\"Ident\":\"uint8\",\"Kind\":8,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"RoundCeil\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"RoundDown\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"RoundFloor\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"RoundUp\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Scan\":{\"In\":[{\"Name\":\"\",\"Ident\":\"interface {}\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"Shift\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Sign\":{\"In\":[],\"Out\":[{\"Name\":\"int\",\"Ident\":\"int\",\"Kind\":2,\"PkgPath\":\"\",\"Methods\":null}]},\"Sin\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"String\":{\"In\":[],\"Out\":[{\"Name\":\"string\",\"Ident\":\"string\",\"Kind\":24,\"PkgPath\":\"\",\"Methods\":null}]},\"StringFixed\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"string\",\"Ident\":\"string\",\"Kind\":24,\"PkgPath\":\"\",\"Methods\":null}]},\"StringFixedBank\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"string\",\"Ident\":\"string\",\"Kind\":24,\"PkgPath\":\"\",\"Methods\":null}]},\"StringFixedCash\":{\"In\":[{\"Name\":\"uint8\",\"Ident\":\"uint8\",\"Kind\":8,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"string\",\"Ident\":\"string\",\"Kind\":24,\"PkgPath\":\"\",\"Methods\":null}]},\"StringScaled\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"string\",\"Ident\":\"string\",\"Kind\":24,\"PkgPath\":\"\",\"Methods\":null}]},\"Sub\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Tan\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Truncate\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"UnmarshalBinary\":{\"In\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"UnmarshalJSON\":{\"In\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"UnmarshalText\":{\"In\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"Value\":{\"In\":[],\"Out\":[{\"Name\":\"Value\",\"Ident\":\"driver.Value\",\"Kind\":20,\"PkgPath\":\"database/sql/driver\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]}}}},\"position\":{\"Index\":7,\"MixedIn\":false,\"MixinIndex\":0},\"schema_type\":{\"mysql\":\"decimal(36,18)\",\"postgres\":\"numeric(36,18)\"},\"annotations\":{\"EntGQL\":{\"Type\":\"String\"}},\"comment\":\"The unit amount of share or money in this account\"},{\"name\":\"value\",\"type\":{\"Type\":20,\"Ident\":\"decimal.Decimal\",\"PkgPath\":\"github.com/shopspring/decimal\",\"PkgName\":\"decimal\",\"Nillable\":false,\"RType\":{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":{\"Abs\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Add\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Atan\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"BigFloat\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"*big.Float\",\"Kind\":22,\"PkgPath\":\"\",\"Methods\":null}]},\"BigInt\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"*big.Int\",\"Kind\":22,\"PkgPath\":\"\",\"Methods\":null}]},\"Ceil\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Cmp\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"int\",\"Ident\":\"int\",\"Kind\":2,\"PkgPath\":\"\",\"Methods\":null}]},\"Coefficient\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"*big.Int\",\"Kind\":22,\"PkgPath\":\"\",\"Methods\":null}]},\"CoefficientInt64\":{\"In\":[],\"Out\":[{\"Name\":\"int64\",\"Ident\":\"int64\",\"Kind\":6,\"PkgPath\":\"\",\"Methods\":null}]},\"Compare\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"int\",\"Ident\":\"int\",\"Kind\":2,\"PkgPath\":\"\",\"Methods\":null}]},\"Copy\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Cos\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Div\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"DivRound\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Equal\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"Equals\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"ExpHullAbrham\":{\"In\":[{\"Name\":\"uint32\",\"Ident\":\"uint32\",\"Kind\":10,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"ExpTaylor\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"Exponent\":{\"In\":[],\"Out\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}]},\"Float64\":{\"In\":[],\"Out\":[{\"Name\":\"float64\",\"Ident\":\"float64\",\"Kind\":14,\"PkgPath\":\"\",\"Methods\":null},{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"Floor\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"GobDecode\":{\"In\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"GobEncode\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"GreaterThan\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"GreaterThanOrEqual\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"InexactFloat64\":{\"In\":[],\"Out\":[{\"Name\":\"float64\",\"Ident\":\"float64\",\"Kind\":14,\"PkgPath\":\"\",\"Methods\":null}]},\"IntPart\":{\"In\":[],\"Out\":[{\"Name\":\"int64\",\"Ident\":\"int64\",\"Kind\":6,\"PkgPath\":\"\",\"Methods\":null}]},\"IsInteger\":{\"In\":[],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"IsNegative\":{\"In\":[],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"IsPositive\":{\"In\":[],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"IsZero\":{\"In\":[],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"LessThan\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"LessThanOrEqual\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"Ln\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"MarshalBinary\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"MarshalJSON\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"MarshalText\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"Mod\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Mul\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Neg\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"NumDigits\":{\"In\":[],\"Out\":[{\"Name\":\"int\",\"Ident\":\"int\",\"Kind\":2,\"PkgPath\":\"\",\"Methods\":null}]},\"Pow\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"PowBigInt\":{\"In\":[{\"Name\":\"\",\"Ident\":\"*big.Int\",\"Kind\":22,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"PowInt32\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"PowWithPrecision\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"QuoRem\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Rat\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"*big.Rat\",\"Kind\":22,\"PkgPath\":\"\",\"Methods\":null}]},\"Round\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"RoundBank\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"RoundCash\":{\"In\":[{\"Name\":\"uint8\",\"Ident\":\"uint8\",\"Kind\":8,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"RoundCeil\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"RoundDown\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"RoundFloor\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"RoundUp\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Scan\":{\"In\":[{\"Name\":\"\",\"Ident\":\"interface {}\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"Shift\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Sign\":{\"In\":[],\"Out\":[{\"Name\":\"int\",\"Ident\":\"int\",\"Kind\":2,\"PkgPath\":\"\",\"Methods\":null}]},\"Sin\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"String\":{\"In\":[],\"Out\":[{\"Name\":\"string\",\"Ident\":\"string\",\"Kind\":24,\"PkgPath\":\"\",\"Methods\":null}]},\"StringFixed\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"string\",\"Ident\":\"string\",\"Kind\":24,\"PkgPath\":\"\",\"Methods\":null}]},\"StringFixedBank\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"string\",\"Ident\":\"string\",\"Kind\":24,\"PkgPath\":\"\",\"Methods\":null}]},\"StringFixedCash\":{\"In\":[{\"Name\":\"uint8\",\"Ident\":\"uint8\",\"Kind\":8,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"string\",\"Ident\":\"string\",\"Kind\":24,\"PkgPath\":\"\",\"Methods\":null}]},\"StringScaled\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"string\",\"Ident\":\"string\",\"Kind\":24,\"PkgPath\":\"\",\"Methods\":null}]},\"Sub\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Tan\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Truncate\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"UnmarshalBinary\":{\"In\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"UnmarshalJSON\":{\"In\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"UnmarshalText\":{\"In\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"Value\":{\"In\":[],\"Out\":[{\"Name\":\"Value\",\"Ident\":\"driver.Value\",\"Kind\":20,\"PkgPath\":\"database/sql/driver\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]}}}},\"position\":{\"Index\":8,\"MixedIn\":false,\"MixinIndex\":0},\"schema_type\":{\"mysql\":\"decimal(18,10)\",\"postgres\":\"numeric(18,10)\"},\"annotations\":{\"EntGQL\":{\"Skip\":48,\"Type\":\"String\"}},\"comment\":\"The value of 1 share in the native currency. If this is just a currency account, then this field will be 1\"},{\"name\":\"balance\",\"type\":{\"Type\":20,\"Ident\":\"decimal.Decimal\",\"PkgPath\":\"github.com/shopspring/decimal\",\"PkgName\":\"decimal\",\"Nillable\":false,\"RType\":{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":{\"Abs\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Add\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Atan\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"BigFloat\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"*big.Float\",\"Kind\":22,\"PkgPath\":\"\",\"Methods\":null}]},\"BigInt\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"*big.Int\",\"Kind\":22,\"PkgPath\":\"\",\"Methods\":null}]},\"Ceil\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Cmp\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"int\",\"Ident\":\"int\",\"Kind\":2,\"PkgPath\":\"\",\"Methods\":null}]},\"Coefficient\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"*big.Int\",\"Kind\":22,\"PkgPath\":\"\",\"Methods\":null}]},\"CoefficientInt64\":{\"In\":[],\"Out\":[{\"Name\":\"int64\",\"Ident\":\"int64\",\"Kind\":6,\"PkgPath\":\"\",\"Methods\":null}]},\"Compare\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"int\",\"Ident\":\"int\",\"Kind\":2,\"PkgPath\":\"\",\"Methods\":null}]},\"Copy\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Cos\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Div\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"DivRound\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Equal\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"Equals\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"ExpHullAbrham\":{\"In\":[{\"Name\":\"uint32\",\"Ident\":\"uint32\",\"Kind\":10,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"ExpTaylor\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"Exponent\":{\"In\":[],\"Out\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}]},\"Float64\":{\"In\":[],\"Out\":[{\"Name\":\"float64\",\"Ident\":\"float64\",\"Kind\":14,\"PkgPath\":\"\",\"Methods\":null},{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"Floor\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"GobDecode\":{\"In\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"GobEncode\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"GreaterThan\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"GreaterThanOrEqual\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"InexactFloat64\":{\"In\":[],\"Out\":[{\"Name\":\"float64\",\"Ident\":\"float64\",\"Kind\":14,\"PkgPath\":\"\",\"Methods\":null}]},\"IntPart\":{\"In\":[],\"Out\":[{\"Name\":\"int64\",\"Ident\":\"int64\",\"Kind\":6,\"PkgPath\":\"\",\"Methods\":null}]},\"IsInteger\":{\"In\":[],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"IsNegative\":{\"In\":[],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"IsPositive\":{\"In\":[],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"IsZero\":{\"In\":[],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"LessThan\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"LessThanOrEqual\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"Ln\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"MarshalBinary\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"MarshalJSON\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"MarshalText\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"Mod\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Mul\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Neg\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"NumDigits\":{\"In\":[],\"Out\":[{\"Name\":\"int\",\"Ident\":\"int\",\"Kind\":2,\"PkgPath\":\"\",\"Methods\":null}]},\"Pow\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"PowBigInt\":{\"In\":[{\"Name\":\"\",\"Ident\":\"*big.Int\",\"Kind\":22,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"PowInt32\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"PowWithPrecision\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"QuoRem\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Rat\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"*big.Rat\",\"Kind\":22,\"PkgPath\":\"\",\"Methods\":null}]},\"Round\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"RoundBank\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"RoundCash\":{\"In\":[{\"Name\":\"uint8\",\"Ident\":\"uint8\",\"Kind\":8,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"RoundCeil\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"RoundDown\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"RoundFloor\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"RoundUp\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Scan\":{\"In\":[{\"Name\":\"\",\"Ident\":\"interface {}\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"Shift\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Sign\":{\"In\":[],\"Out\":[{\"Name\":\"int\",\"Ident\":\"int\",\"Kind\":2,\"PkgPath\":\"\",\"Methods\":null}]},\"Sin\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"String\":{\"In\":[],\"Out\":[{\"Name\":\"string\",\"Ident\":\"string\",\"Kind\":24,\"PkgPath\":\"\",\"Methods\":null}]},\"StringFixed\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"string\",\"Ident\":\"string\",\"Kind\":24,\"PkgPath\":\"\",\"Methods\":null}]},\"StringFixedBank\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"string\",\"Ident\":\"string\",\"Kind\":24,\"PkgPath\":\"\",\"Methods\":null}]},\"StringFixedCash\":{\"In\":[{\"Name\":\"uint8\",\"Ident\":\"uint8\",\"Kind\":8,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"string\",\"Ident\":\"string\",\"Kind\":24,\"PkgPath\":\"\",\"Methods\":null}]},\"StringScaled\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"string\",\"Ident\":\"string\",\"Kind\":24,\"PkgPath\":\"\",\"Methods\":null}]},\"Sub\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Tan\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Truncate\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"UnmarshalBinary\":{\"In\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"UnmarshalJSON\":{\"In\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"UnmarshalText\":{\"In\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"Value\":{\"In\":[],\"Out\":[{\"Name\":\"Value\",\"Ident\":\"driver.Value\",\"Kind\":20,\"PkgPath\":\"database/sql/driver\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]}}}},\"position\":{\"Index\":9,\"MixedIn\":false,\"MixinIndex\":0},\"schema_type\":{\"mysql\":\"decimal(36,18)\",\"postgres\":\"numeric(36,18)\"},\"annotations\":{\"EntGQL\":{\"Skip\":48,\"Type\":\"String\"}},\"comment\":\"The total balance of this account in its currency\"},{\"name\":\"archived\",\"type\":{\"Type\":1,\"Ident\":\"\",\"PkgPath\":\"\",\"PkgName\":\"\",\"Nillable\":false,\"RType\":null},\"default\":true,\"default_value\":false,\"default_kind\":1,\"position\":{\"Index\":10,\"MixedIn\":false,\"MixinIndex\":0}}],\"annotations\":{\"EntGQL\":{\"MutationInputs\":[{\"IsCreate\":true},{}],\"QueryField\":{},\"RelayConnection\":true},\"EntSQL\":{\"increment_start\":0}}},{\"name\":\"Category\",\"config\":{\"Table\":\"\"},\"edges\":[{\"name\":\"profile\",\"type\":\"Profile\",\"ref_name\":\"categories\",\"unique\":true,\"inverse\":true,\"required\":true,\"annotations\":{\"EntGQL\":{\"Skip\":48}}},{\"name\":\"transactions\",\"type\":\"Transaction\"}],\"fields\":[{\"name\":\"create_time\",\"type\":{\"Type\":2,\"Ident\":\"\",\"PkgPath\":\"time\",\"PkgName\":\"\",\"Nillable\":false,\"RType\":null},\"default\":true,\"default_kind\":19,\"immutable\":true,\"position\":{\"Index\":0,\"MixedIn\":true,\"MixinIndex\":0}},{\"name\":\"update_time\",\"type\":{\"Type\":2,\"Ident\":\"\",\"PkgPath\":\"time\",\"PkgName\":\"\",\"Nillable\":false,\"RType\":null},\"default\":true,\"default_kind\":19,\"update_default\":true,\"position\":{\"Index\":1,\"MixedIn\":true,\"MixinIndex\":0}},{\"name\":\"name\",\"type\":{\"Type\":7,\"Ident\":\"\",\"PkgPath\":\"\",\"PkgName\":\"\",\"Nillable\":false,\"RType\":null},\"validators\":1,\"position\":{\"Index\":0,\"MixedIn\":false,\"MixinIndex\":0}},{\"name\":\"category_type\",\"type\":{\"Type\":6,\"Ident\":\"category.CategoryType\",\"PkgPath\":\"\",\"PkgName\":\"\",\"Nillable\":false,\"RType\":null},\"enums\":[{\"N\":\"expense\",\"V\":\"expense\"},{\"N\":\"income\",\"V\":\"income\"},{\"N\":\"transfer\",\"V\":\"transfer\"},{\"N\":\"sync\",\"V\":\"sync\"},{\"N\":\"init\",\"V\":\"init\"}],\"immutable\":true,\"position\":{\"Index\":1,\"MixedIn\":false,\"MixinIndex\":0}}],\"annotations\":{\"EntGQL\":{\"MutationInputs\":[{\"IsCreate\":true},{}],\"QueryField\":{}},\"EntSQL\":{\"increment_start\":38654705664}}},{\"name\":\"Profile\",\"config\":{\"Table\":\"\"},\"edges\":[{\"name\":\"users\",\"type\":\"User\",\"ref_name\":\"profiles\",\"through\":{\"N\":\"user_profiles\",\"T\":\"UserProfile\"},\"inverse\":true,\"annotations\":{\"EntGQL\":{\"Skip\":48}}},{\"name\":\"accounts\",\"type\":\"Account\",\"annotations\":{\"EntSQL\":{\"on_delete\":\"CASCADE\"}}},{\"name\":\"transactions\",\"type\":\"Transaction\",\"annotations\":{\"EntSQL\":{\"on_delete\":\"CASCADE\"}}},{\"name\":\"snapshots\",\"type\":\"Snapshot\",\"annotations\":{\"EntSQL\":{\"on_delete\":\"CASCADE\"}}},{\"name\":\"categories\",\"type\":\"Category\",\"annotations\":{\"EntSQL\":{\"on_delete\":\"CASCADE\"}}}],\"fields\":[{\"name\":\"create_time\",\"type\":{\"Type\":2,\"Ident\":\"\",\"PkgPath\":\"time\",\"PkgName\":\"\",\"Nillable\":false,\"RType\":null},\"default\":true,\"default_kind\":19,\"immutable\":true,\"position\":{\"Index\":0,\"MixedIn\":true,\"MixinIndex\":0}},{\"name\":\"update_time\",\"type\":{\"Type\":2,\"Ident\":\"\",\"PkgPath\":\"time\",\"PkgName\":\"\",\"Nillable\":false,\"RType\":null},\"default\":true,\"default_kind\":19,\"update_default\":true,\"position\":{\"Index\":1,\"MixedIn\":true,\"MixinIndex\":0}},{\"name\":\"name\",\"type\":{\"Type\":7,\"Ident\":\"\",\"PkgPath\":\"\",\"PkgName\":\"\",\"Nillable\":false,\"RType\":null},\"position\":{\"Index\":0,\"MixedIn\":false,\"MixinIndex\":0}},{\"name\":\"locale\",\"type\":{\"Type\":7,\"Ident\":\"\",\"PkgPath\":\"\",\"PkgName\":\"\",\"Nillable\":false,\"RType\":null},\"position\":{\"Index\":1,\"MixedIn\":false,\"MixinIndex\":0},\"annotations\":{\"EntGQL\":{\"Skip\":48}}},{\"name\":\"currencies\",\"type\":{\"Type\":3,\"Ident\":\"[]string\",\"PkgPath\":\"\",\"PkgName\":\"\",\"Nillable\":true,\"RType\":{\"Name\":\"\",\"Ident\":\"[]string\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":{}}},\"position\":{\"Index\":2,\"MixedIn\":false,\"MixinIndex\":0}},{\"name\":\"net_worth_goal\",\"type\":{\"Type\":20,\"Ident\":\"decimal.Decimal\",\"PkgPath\":\"github.com/shopspring/decimal\",\"PkgName\":\"decimal\",\"Nillable\":false,\"RType\":{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":{\"Abs\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Add\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Atan\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"BigFloat\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"*big.Float\",\"Kind\":22,\"PkgPath\":\"\",\"Methods\":null}]},\"BigInt\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"*big.Int\",\"Kind\":22,\"PkgPath\":\"\",\"Methods\":null}]},\"Ceil\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Cmp\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"int\",\"Ident\":\"int\",\"Kind\":2,\"PkgPath\":\"\",\"Methods\":null}]},\"Coefficient\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"*big.Int\",\"Kind\":22,\"PkgPath\":\"\",\"Methods\":null}]},\"CoefficientInt64\":{\"In\":[],\"Out\":[{\"Name\":\"int64\",\"Ident\":\"int64\",\"Kind\":6,\"PkgPath\":\"\",\"Methods\":null}]},\"Compare\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"int\",\"Ident\":\"int\",\"Kind\":2,\"PkgPath\":\"\",\"Methods\":null}]},\"Copy\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Cos\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Div\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"DivRound\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Equal\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"Equals\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"ExpHullAbrham\":{\"In\":[{\"Name\":\"uint32\",\"Ident\":\"uint32\",\"Kind\":10,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"ExpTaylor\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"Exponent\":{\"In\":[],\"Out\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}]},\"Float64\":{\"In\":[],\"Out\":[{\"Name\":\"float64\",\"Ident\":\"float64\",\"Kind\":14,\"PkgPath\":\"\",\"Methods\":null},{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"Floor\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"GobDecode\":{\"In\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"GobEncode\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"GreaterThan\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"GreaterThanOrEqual\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"InexactFloat64\":{\"In\":[],\"Out\":[{\"Name\":\"float64\",\"Ident\":\"float64\",\"Kind\":14,\"PkgPath\":\"\",\"Methods\":null}]},\"IntPart\":{\"In\":[],\"Out\":[{\"Name\":\"int64\",\"Ident\":\"int64\",\"Kind\":6,\"PkgPath\":\"\",\"Methods\":null}]},\"IsInteger\":{\"In\":[],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"IsNegative\":{\"In\":[],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"IsPositive\":{\"In\":[],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"IsZero\":{\"In\":[],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"LessThan\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"LessThanOrEqual\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"Ln\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"MarshalBinary\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"MarshalJSON\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"MarshalText\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"Mod\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Mul\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Neg\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"NumDigits\":{\"In\":[],\"Out\":[{\"Name\":\"int\",\"Ident\":\"int\",\"Kind\":2,\"PkgPath\":\"\",\"Methods\":null}]},\"Pow\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"PowBigInt\":{\"In\":[{\"Name\":\"\",\"Ident\":\"*big.Int\",\"Kind\":22,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"PowInt32\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"PowWithPrecision\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"QuoRem\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Rat\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"*big.Rat\",\"Kind\":22,\"PkgPath\":\"\",\"Methods\":null}]},\"Round\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"RoundBank\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"RoundCash\":{\"In\":[{\"Name\":\"uint8\",\"Ident\":\"uint8\",\"Kind\":8,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"RoundCeil\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"RoundDown\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"RoundFloor\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"RoundUp\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Scan\":{\"In\":[{\"Name\":\"\",\"Ident\":\"interface {}\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"Shift\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Sign\":{\"In\":[],\"Out\":[{\"Name\":\"int\",\"Ident\":\"int\",\"Kind\":2,\"PkgPath\":\"\",\"Methods\":null}]},\"Sin\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"String\":{\"In\":[],\"Out\":[{\"Name\":\"string\",\"Ident\":\"string\",\"Kind\":24,\"PkgPath\":\"\",\"Methods\":null}]},\"StringFixed\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"string\",\"Ident\":\"string\",\"Kind\":24,\"PkgPath\":\"\",\"Methods\":null}]},\"StringFixedBank\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"string\",\"Ident\":\"string\",\"Kind\":24,\"PkgPath\":\"\",\"Methods\":null}]},\"StringFixedCash\":{\"In\":[{\"Name\":\"uint8\",\"Ident\":\"uint8\",\"Kind\":8,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"string\",\"Ident\":\"string\",\"Kind\":24,\"PkgPath\":\"\",\"Methods\":null}]},\"StringScaled\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"string\",\"Ident\":\"string\",\"Kind\":24,\"PkgPath\":\"\",\"Methods\":null}]},\"Sub\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Tan\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Truncate\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"UnmarshalBinary\":{\"In\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"UnmarshalJSON\":{\"In\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"UnmarshalText\":{\"In\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"Value\":{\"In\":[],\"Out\":[{\"Name\":\"Value\",\"Ident\":\"driver.Value\",\"Kind\":20,\"PkgPath\":\"database/sql/driver\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]}}}},\"position\":{\"Index\":3,\"MixedIn\":false,\"MixinIndex\":0},\"schema_type\":{\"mysql\":\"decimal(36,18)\",\"postgres\":\"numeric(36,18)\"},\"annotations\":{\"EntGQL\":{\"Type\":\"String\"}}}],\"annotations\":{\"EntGQL\":{\"MutationInputs\":[{\"IsCreate\":true},{}],\"QueryField\":{}},\"EntSQL\":{\"increment_start\":4294967296}}},{\"name\":\"Snapshot\",\"config\":{\"Table\":\"\"},\"edges\":[{\"name\":\"profile\",\"type\":\"Profile\",\"ref_name\":\"snapshots\",\"unique\":true,\"inverse\":true,\"required\":true,\"annotations\":{\"EntGQL\":{\"Skip\":48}}},{\"name\":\"snapshot_accounts\",\"type\":\"SnapshotAccount\",\"annotations\":{\"EntSQL\":{\"on_delete\":\"CASCADE\"}}},{\"name\":\"snapshot_fx_rates\",\"type\":\"SnapshotFXRate\",\"annotations\":{\"EntSQL\":{\"on_delete\":\"CASCADE\"}}}],\"fields\":[{\"name\":\"create_time\",\"type\":{\"Type\":2,\"Ident\":\"\",\"PkgPath\":\"time\",\"PkgName\":\"\",\"Nillable\":false,\"RType\":null},\"default\":true,\"default_kind\":19,\"immutable\":true,\"position\":{\"Index\":0,\"MixedIn\":true,\"MixinIndex\":0}},{\"name\":\"update_time\",\"type\":{\"Type\":2,\"Ident\":\"\",\"PkgPath\":\"time\",\"PkgName\":\"\",\"Nillable\":false,\"RType\":null},\"default\":true,\"default_kind\":19,\"update_default\":true,\"position\":{\"Index\":1,\"MixedIn\":true,\"MixinIndex\":0}},{\"name\":\"note\",\"type\":{\"Type\":7,\"Ident\":\"\",\"PkgPath\":\"\",\"PkgName\":\"\",\"Nillable\":false,\"RType\":null},\"size\":2147483647,\"optional\":true,\"position\":{\"Index\":0,\"MixedIn\":false,\"MixinIndex\":0}}],\"annotations\":{\"EntGQL\":{\"QueryField\":{}},\"EntSQL\":{\"increment_start\":25769803776}}},{\"name\":\"SnapshotAccount\",\"config\":{\"Table\":\"\"},\"edges\":[{\"name\":\"account\",\"type\":\"Account\",\"ref_name\":\"snapshot_accounts\",\"unique\":true,\"inverse\":true,\"required\":true,\"immutable\":true,\"annotations\":{\"EntGQL\":{\"Skip\":48}}},{\"name\":\"snapshot\",\"type\":\"Snapshot\",\"ref_name\":\"snapshot_accounts\",\"unique\":true,\"inverse\":true,\"required\":true,\"immutable\":true,\"annotations\":{\"EntGQL\":{\"Skip\":48}}}],\"fields\":[{\"name\":\"create_time\",\"type\":{\"Type\":2,\"Ident\":\"\",\"PkgPath\":\"time\",\"PkgName\":\"\",\"Nillable\":false,\"RType\":null},\"default\":true,\"default_kind\":19,\"immutable\":true,\"position\":{\"Index\":0,\"MixedIn\":true,\"MixinIndex\":0}},{\"name\":\"update_time\",\"type\":{\"Type\":2,\"Ident\":\"\",\"PkgPath\":\"time\",\"PkgName\":\"\",\"Nillable\":false,\"RType\":null},\"default\":true,\"default_kind\":19,\"update_default\":true,\"position\":{\"Index\":1,\"MixedIn\":true,\"MixinIndex\":0}},{\"name\":\"amount\",\"type\":{\"Type\":20,\"Ident\":\"decimal.Decimal\",\"PkgPath\":\"github.com/shopspring/decimal\",\"PkgName\":\"decimal\",\"Nillable\":false,\"RType\":{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":{\"Abs\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Add\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Atan\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"BigFloat\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"*big.Float\",\"Kind\":22,\"PkgPath\":\"\",\"Methods\":null}]},\"BigInt\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"*big.Int\",\"Kind\":22,\"PkgPath\":\"\",\"Methods\":null}]},\"Ceil\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Cmp\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"int\",\"Ident\":\"int\",\"Kind\":2,\"PkgPath\":\"\",\"Methods\":null}]},\"Coefficient\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"*big.Int\",\"Kind\":22,\"PkgPath\":\"\",\"Methods\":null}]},\"CoefficientInt64\":{\"In\":[],\"Out\":[{\"Name\":\"int64\",\"Ident\":\"int64\",\"Kind\":6,\"PkgPath\":\"\",\"Methods\":null}]},\"Compare\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"int\",\"Ident\":\"int\",\"Kind\":2,\"PkgPath\":\"\",\"Methods\":null}]},\"Copy\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Cos\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Div\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"DivRound\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Equal\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"Equals\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"ExpHullAbrham\":{\"In\":[{\"Name\":\"uint32\",\"Ident\":\"uint32\",\"Kind\":10,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"ExpTaylor\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"Exponent\":{\"In\":[],\"Out\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}]},\"Float64\":{\"In\":[],\"Out\":[{\"Name\":\"float64\",\"Ident\":\"float64\",\"Kind\":14,\"PkgPath\":\"\",\"Methods\":null},{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"Floor\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"GobDecode\":{\"In\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"GobEncode\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"GreaterThan\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"GreaterThanOrEqual\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"InexactFloat64\":{\"In\":[],\"Out\":[{\"Name\":\"float64\",\"Ident\":\"float64\",\"Kind\":14,\"PkgPath\":\"\",\"Methods\":null}]},\"IntPart\":{\"In\":[],\"Out\":[{\"Name\":\"int64\",\"Ident\":\"int64\",\"Kind\":6,\"PkgPath\":\"\",\"Methods\":null}]},\"IsInteger\":{\"In\":[],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"IsNegative\":{\"In\":[],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"IsPositive\":{\"In\":[],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"IsZero\":{\"In\":[],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"LessThan\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"LessThanOrEqual\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"Ln\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"MarshalBinary\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"MarshalJSON\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"MarshalText\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"Mod\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Mul\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Neg\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"NumDigits\":{\"In\":[],\"Out\":[{\"Name\":\"int\",\"Ident\":\"int\",\"Kind\":2,\"PkgPath\":\"\",\"Methods\":null}]},\"Pow\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"PowBigInt\":{\"In\":[{\"Name\":\"\",\"Ident\":\"*big.Int\",\"Kind\":22,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"PowInt32\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"PowWithPrecision\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"QuoRem\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Rat\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"*big.Rat\",\"Kind\":22,\"PkgPath\":\"\",\"Methods\":null}]},\"Round\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"RoundBank\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"RoundCash\":{\"In\":[{\"Name\":\"uint8\",\"Ident\":\"uint8\",\"Kind\":8,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"RoundCeil\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"RoundDown\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"RoundFloor\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"RoundUp\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Scan\":{\"In\":[{\"Name\":\"\",\"Ident\":\"interface {}\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"Shift\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Sign\":{\"In\":[],\"Out\":[{\"Name\":\"int\",\"Ident\":\"int\",\"Kind\":2,\"PkgPath\":\"\",\"Methods\":null}]},\"Sin\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"String\":{\"In\":[],\"Out\":[{\"Name\":\"string\",\"Ident\":\"string\",\"Kind\":24,\"PkgPath\":\"\",\"Methods\":null}]},\"StringFixed\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"string\",\"Ident\":\"string\",\"Kind\":24,\"PkgPath\":\"\",\"Methods\":null}]},\"StringFixedBank\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"string\",\"Ident\":\"string\",\"Kind\":24,\"PkgPath\":\"\",\"Methods\":null}]},\"StringFixedCash\":{\"In\":[{\"Name\":\"uint8\",\"Ident\":\"uint8\",\"Kind\":8,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"string\",\"Ident\":\"string\",\"Kind\":24,\"PkgPath\":\"\",\"Methods\":null}]},\"StringScaled\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"string\",\"Ident\":\"string\",\"Kind\":24,\"PkgPath\":\"\",\"Methods\":null}]},\"Sub\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Tan\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Truncate\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"UnmarshalBinary\":{\"In\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"UnmarshalJSON\":{\"In\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"UnmarshalText\":{\"In\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"Value\":{\"In\":[],\"Out\":[{\"Name\":\"Value\",\"Ident\":\"driver.Value\",\"Kind\":20,\"PkgPath\":\"database/sql/driver\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]}}}},\"immutable\":true,\"position\":{\"Index\":0,\"MixedIn\":false,\"MixinIndex\":0},\"schema_type\":{\"mysql\":\"decimal(36,18)\",\"postgres\":\"numeric(36,18)\"},\"annotations\":{\"EntGQL\":{\"Type\":\"String\"}},\"comment\":\"The unit amount of share or money in this account\"},{\"name\":\"value\",\"type\":{\"Type\":20,\"Ident\":\"decimal.Decimal\",\"PkgPath\":\"github.com/shopspring/decimal\",\"PkgName\":\"decimal\",\"Nillable\":false,\"RType\":{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":{\"Abs\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Add\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Atan\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"BigFloat\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"*big.Float\",\"Kind\":22,\"PkgPath\":\"\",\"Methods\":null}]},\"BigInt\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"*big.Int\",\"Kind\":22,\"PkgPath\":\"\",\"Methods\":null}]},\"Ceil\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Cmp\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"int\",\"Ident\":\"int\",\"Kind\":2,\"PkgPath\":\"\",\"Methods\":null}]},\"Coefficient\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"*big.Int\",\"Kind\":22,\"PkgPath\":\"\",\"Methods\":null}]},\"CoefficientInt64\":{\"In\":[],\"Out\":[{\"Name\":\"int64\",\"Ident\":\"int64\",\"Kind\":6,\"PkgPath\":\"\",\"Methods\":null}]},\"Compare\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"int\",\"Ident\":\"int\",\"Kind\":2,\"PkgPath\":\"\",\"Methods\":null}]},\"Copy\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Cos\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Div\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"DivRound\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Equal\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"Equals\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"ExpHullAbrham\":{\"In\":[{\"Name\":\"uint32\",\"Ident\":\"uint32\",\"Kind\":10,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"ExpTaylor\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"Exponent\":{\"In\":[],\"Out\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}]},\"Float64\":{\"In\":[],\"Out\":[{\"Name\":\"float64\",\"Ident\":\"float64\",\"Kind\":14,\"PkgPath\":\"\",\"Methods\":null},{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"Floor\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"GobDecode\":{\"In\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"GobEncode\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"GreaterThan\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"GreaterThanOrEqual\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"InexactFloat64\":{\"In\":[],\"Out\":[{\"Name\":\"float64\",\"Ident\":\"float64\",\"Kind\":14,\"PkgPath\":\"\",\"Methods\":null}]},\"IntPart\":{\"In\":[],\"Out\":[{\"Name\":\"int64\",\"Ident\":\"int64\",\"Kind\":6,\"PkgPath\":\"\",\"Methods\":null}]},\"IsInteger\":{\"In\":[],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"IsNegative\":{\"In\":[],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"IsPositive\":{\"In\":[],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"IsZero\":{\"In\":[],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"LessThan\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"LessThanOrEqual\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"Ln\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"MarshalBinary\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"MarshalJSON\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"MarshalText\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"Mod\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Mul\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Neg\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"NumDigits\":{\"In\":[],\"Out\":[{\"Name\":\"int\",\"Ident\":\"int\",\"Kind\":2,\"PkgPath\":\"\",\"Methods\":null}]},\"Pow\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"PowBigInt\":{\"In\":[{\"Name\":\"\",\"Ident\":\"*big.Int\",\"Kind\":22,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"PowInt32\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"PowWithPrecision\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"QuoRem\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Rat\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"*big.Rat\",\"Kind\":22,\"PkgPath\":\"\",\"Methods\":null}]},\"Round\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"RoundBank\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"RoundCash\":{\"In\":[{\"Name\":\"uint8\",\"Ident\":\"uint8\",\"Kind\":8,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"RoundCeil\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"RoundDown\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"RoundFloor\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"RoundUp\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Scan\":{\"In\":[{\"Name\":\"\",\"Ident\":\"interface {}\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"Shift\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Sign\":{\"In\":[],\"Out\":[{\"Name\":\"int\",\"Ident\":\"int\",\"Kind\":2,\"PkgPath\":\"\",\"Methods\":null}]},\"Sin\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"String\":{\"In\":[],\"Out\":[{\"Name\":\"string\",\"Ident\":\"string\",\"Kind\":24,\"PkgPath\":\"\",\"Methods\":null}]},\"StringFixed\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"string\",\"Ident\":\"string\",\"Kind\":24,\"PkgPath\":\"\",\"Methods\":null}]},\"StringFixedBank\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"string\",\"Ident\":\"string\",\"Kind\":24,\"PkgPath\":\"\",\"Methods\":null}]},\"StringFixedCash\":{\"In\":[{\"Name\":\"uint8\",\"Ident\":\"uint8\",\"Kind\":8,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"string\",\"Ident\":\"string\",\"Kind\":24,\"PkgPath\":\"\",\"Methods\":null}]},\"StringScaled\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"string\",\"Ident\":\"string\",\"Kind\":24,\"PkgPath\":\"\",\"Methods\":null}]},\"Sub\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Tan\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Truncate\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"UnmarshalBinary\":{\"In\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"UnmarshalJSON\":{\"In\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"UnmarshalText\":{\"In\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"Value\":{\"In\":[],\"Out\":[{\"Name\":\"Value\",\"Ident\":\"driver.Value\",\"Kind\":20,\"PkgPath\":\"database/sql/driver\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]}}}},\"immutable\":true,\"position\":{\"Index\":1,\"MixedIn\":false,\"MixinIndex\":0},\"schema_type\":{\"mysql\":\"decimal(18,10)\",\"postgres\":\"numeric(18,10)\"},\"annotations\":{\"EntGQL\":{\"Skip\":48,\"Type\":\"String\"}},\"comment\":\"The value of 1 share in the native currency. If this is just a currency account, then this field will be 1\"},{\"name\":\"balance\",\"type\":{\"Type\":20,\"Ident\":\"decimal.Decimal\",\"PkgPath\":\"github.com/shopspring/decimal\",\"PkgName\":\"decimal\",\"Nillable\":false,\"RType\":{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":{\"Abs\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Add\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Atan\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"BigFloat\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"*big.Float\",\"Kind\":22,\"PkgPath\":\"\",\"Methods\":null}]},\"BigInt\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"*big.Int\",\"Kind\":22,\"PkgPath\":\"\",\"Methods\":null}]},\"Ceil\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Cmp\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"int\",\"Ident\":\"int\",\"Kind\":2,\"PkgPath\":\"\",\"Methods\":null}]},\"Coefficient\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"*big.Int\",\"Kind\":22,\"PkgPath\":\"\",\"Methods\":null}]},\"CoefficientInt64\":{\"In\":[],\"Out\":[{\"Name\":\"int64\",\"Ident\":\"int64\",\"Kind\":6,\"PkgPath\":\"\",\"Methods\":null}]},\"Compare\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"int\",\"Ident\":\"int\",\"Kind\":2,\"PkgPath\":\"\",\"Methods\":null}]},\"Copy\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Cos\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Div\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"DivRound\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Equal\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"Equals\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"ExpHullAbrham\":{\"In\":[{\"Name\":\"uint32\",\"Ident\":\"uint32\",\"Kind\":10,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"ExpTaylor\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"Exponent\":{\"In\":[],\"Out\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}]},\"Float64\":{\"In\":[],\"Out\":[{\"Name\":\"float64\",\"Ident\":\"float64\",\"Kind\":14,\"PkgPath\":\"\",\"Methods\":null},{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"Floor\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"GobDecode\":{\"In\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"GobEncode\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"GreaterThan\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"GreaterThanOrEqual\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"InexactFloat64\":{\"In\":[],\"Out\":[{\"Name\":\"float64\",\"Ident\":\"float64\",\"Kind\":14,\"PkgPath\":\"\",\"Methods\":null}]},\"IntPart\":{\"In\":[],\"Out\":[{\"Name\":\"int64\",\"Ident\":\"int64\",\"Kind\":6,\"PkgPath\":\"\",\"Methods\":null}]},\"IsInteger\":{\"In\":[],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"IsNegative\":{\"In\":[],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"IsPositive\":{\"In\":[],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"IsZero\":{\"In\":[],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"LessThan\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"LessThanOrEqual\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"Ln\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"MarshalBinary\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"MarshalJSON\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"MarshalText\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"Mod\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Mul\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Neg\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"NumDigits\":{\"In\":[],\"Out\":[{\"Name\":\"int\",\"Ident\":\"int\",\"Kind\":2,\"PkgPath\":\"\",\"Methods\":null}]},\"Pow\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"PowBigInt\":{\"In\":[{\"Name\":\"\",\"Ident\":\"*big.Int\",\"Kind\":22,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"PowInt32\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"PowWithPrecision\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"QuoRem\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Rat\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"*big.Rat\",\"Kind\":22,\"PkgPath\":\"\",\"Methods\":null}]},\"Round\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"RoundBank\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"RoundCash\":{\"In\":[{\"Name\":\"uint8\",\"Ident\":\"uint8\",\"Kind\":8,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"RoundCeil\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"RoundDown\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"RoundFloor\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"RoundUp\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Scan\":{\"In\":[{\"Name\":\"\",\"Ident\":\"interface {}\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"Shift\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Sign\":{\"In\":[],\"Out\":[{\"Name\":\"int\",\"Ident\":\"int\",\"Kind\":2,\"PkgPath\":\"\",\"Methods\":null}]},\"Sin\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"String\":{\"In\":[],\"Out\":[{\"Name\":\"string\",\"Ident\":\"string\",\"Kind\":24,\"PkgPath\":\"\",\"Methods\":null}]},\"StringFixed\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"string\",\"Ident\":\"string\",\"Kind\":24,\"PkgPath\":\"\",\"Methods\":null}]},\"StringFixedBank\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"string\",\"Ident\":\"string\",\"Kind\":24,\"PkgPath\":\"\",\"Methods\":null}]},\"StringFixedCash\":{\"In\":[{\"Name\":\"uint8\",\"Ident\":\"uint8\",\"Kind\":8,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"string\",\"Ident\":\"string\",\"Kind\":24,\"PkgPath\":\"\",\"Methods\":null}]},\"StringScaled\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"string\",\"Ident\":\"string\",\"Kind\":24,\"PkgPath\":\"\",\"Methods\":null}]},\"Sub\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Tan\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Truncate\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"UnmarshalBinary\":{\"In\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"UnmarshalJSON\":{\"In\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"UnmarshalText\":{\"In\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"Value\":{\"In\":[],\"Out\":[{\"Name\":\"Value\",\"Ident\":\"driver.Value\",\"Kind\":20,\"PkgPath\":\"database/sql/driver\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]}}}},\"immutable\":true,\"position\":{\"Index\":2,\"MixedIn\":false,\"MixinIndex\":0},\"schema_type\":{\"mysql\":\"decimal(36,18)\",\"postgres\":\"numeric(36,18)\"},\"annotations\":{\"EntGQL\":{\"Skip\":48,\"Type\":\"String\"}},\"comment\":\"The total balance of this account its currency\"}],\"annotations\":{\"EntGQL\":{\"QueryField\":{}},\"EntSQL\":{\"increment_start\":30064771072}}},{\"name\":\"SnapshotFXRate\",\"config\":{\"Table\":\"\"},\"edges\":[{\"name\":\"snapshot\",\"type\":\"Snapshot\",\"ref_name\":\"snapshot_fx_rates\",\"unique\":true,\"inverse\":true,\"required\":true,\"immutable\":true,\"annotations\":{\"EntGQL\":{\"Skip\":48}}}],\"fields\":[{\"name\":\"create_time\",\"type\":{\"Type\":2,\"Ident\":\"\",\"PkgPath\":\"time\",\"PkgName\":\"\",\"Nillable\":false,\"RType\":null},\"default\":true,\"default_kind\":19,\"immutable\":true,\"position\":{\"Index\":0,\"MixedIn\":true,\"MixinIndex\":0}},{\"name\":\"update_time\",\"type\":{\"Type\":2,\"Ident\":\"\",\"PkgPath\":\"time\",\"PkgName\":\"\",\"Nillable\":false,\"RType\":null},\"default\":true,\"default_kind\":19,\"update_default\":true,\"position\":{\"Index\":1,\"MixedIn\":true,\"MixinIndex\":0}},{\"name\":\"from_currency\",\"type\":{\"Type\":7,\"Ident\":\"\",\"PkgPath\":\"\",\"PkgName\":\"\",\"Nillable\":false,\"RType\":null},\"size\":2147483647,\"immutable\":true,\"position\":{\"Index\":0,\"MixedIn\":false,\"MixinIndex\":0}},{\"name\":\"to_currency\",\"type\":{\"Type\":7,\"Ident\":\"\",\"PkgPath\":\"\",\"PkgName\":\"\",\"Nillable\":false,\"RType\":null},\"size\":2147483647,\"immutable\":true,\"position\":{\"Index\":1,\"MixedIn\":false,\"MixinIndex\":0}},{\"name\":\"fx_rate\",\"type\":{\"Type\":20,\"Ident\":\"decimal.Decimal\",\"PkgPath\":\"github.com/shopspring/decimal\",\"PkgName\":\"decimal\",\"Nillable\":false,\"RType\":{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":{\"Abs\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Add\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Atan\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"BigFloat\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"*big.Float\",\"Kind\":22,\"PkgPath\":\"\",\"Methods\":null}]},\"BigInt\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"*big.Int\",\"Kind\":22,\"PkgPath\":\"\",\"Methods\":null}]},\"Ceil\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Cmp\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"int\",\"Ident\":\"int\",\"Kind\":2,\"PkgPath\":\"\",\"Methods\":null}]},\"Coefficient\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"*big.Int\",\"Kind\":22,\"PkgPath\":\"\",\"Methods\":null}]},\"CoefficientInt64\":{\"In\":[],\"Out\":[{\"Name\":\"int64\",\"Ident\":\"int64\",\"Kind\":6,\"PkgPath\":\"\",\"Methods\":null}]},\"Compare\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"int\",\"Ident\":\"int\",\"Kind\":2,\"PkgPath\":\"\",\"Methods\":null}]},\"Copy\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Cos\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Div\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"DivRound\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Equal\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"Equals\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"ExpHullAbrham\":{\"In\":[{\"Name\":\"uint32\",\"Ident\":\"uint32\",\"Kind\":10,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"ExpTaylor\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"Exponent\":{\"In\":[],\"Out\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}]},\"Float64\":{\"In\":[],\"Out\":[{\"Name\":\"float64\",\"Ident\":\"float64\",\"Kind\":14,\"PkgPath\":\"\",\"Methods\":null},{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"Floor\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"GobDecode\":{\"In\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"GobEncode\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"GreaterThan\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"GreaterThanOrEqual\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"InexactFloat64\":{\"In\":[],\"Out\":[{\"Name\":\"float64\",\"Ident\":\"float64\",\"Kind\":14,\"PkgPath\":\"\",\"Methods\":null}]},\"IntPart\":{\"In\":[],\"Out\":[{\"Name\":\"int64\",\"Ident\":\"int64\",\"Kind\":6,\"PkgPath\":\"\",\"Methods\":null}]},\"IsInteger\":{\"In\":[],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"IsNegative\":{\"In\":[],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"IsPositive\":{\"In\":[],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"IsZero\":{\"In\":[],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"LessThan\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"LessThanOrEqual\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"Ln\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"MarshalBinary\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"MarshalJSON\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"MarshalText\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"Mod\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Mul\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Neg\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"NumDigits\":{\"In\":[],\"Out\":[{\"Name\":\"int\",\"Ident\":\"int\",\"Kind\":2,\"PkgPath\":\"\",\"Methods\":null}]},\"Pow\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"PowBigInt\":{\"In\":[{\"Name\":\"\",\"Ident\":\"*big.Int\",\"Kind\":22,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"PowInt32\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"PowWithPrecision\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"QuoRem\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Rat\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"*big.Rat\",\"Kind\":22,\"PkgPath\":\"\",\"Methods\":null}]},\"Round\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"RoundBank\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"RoundCash\":{\"In\":[{\"Name\":\"uint8\",\"Ident\":\"uint8\",\"Kind\":8,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"RoundCeil\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"RoundDown\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"RoundFloor\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"RoundUp\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Scan\":{\"In\":[{\"Name\":\"\",\"Ident\":\"interface {}\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"Shift\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Sign\":{\"In\":[],\"Out\":[{\"Name\":\"int\",\"Ident\":\"int\",\"Kind\":2,\"PkgPath\":\"\",\"Methods\":null}]},\"Sin\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"String\":{\"In\":[],\"Out\":[{\"Name\":\"string\",\"Ident\":\"string\",\"Kind\":24,\"PkgPath\":\"\",\"Methods\":null}]},\"StringFixed\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"string\",\"Ident\":\"string\",\"Kind\":24,\"PkgPath\":\"\",\"Methods\":null}]},\"StringFixedBank\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"string\",\"Ident\":\"string\",\"Kind\":24,\"PkgPath\":\"\",\"Methods\":null}]},\"StringFixedCash\":{\"In\":[{\"Name\":\"uint8\",\"Ident\":\"uint8\",\"Kind\":8,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"string\",\"Ident\":\"string\",\"Kind\":24,\"PkgPath\":\"\",\"Methods\":null}]},\"StringScaled\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"string\",\"Ident\":\"string\",\"Kind\":24,\"PkgPath\":\"\",\"Methods\":null}]},\"Sub\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Tan\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Truncate\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"UnmarshalBinary\":{\"In\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"UnmarshalJSON\":{\"In\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"UnmarshalText\":{\"In\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"Value\":{\"In\":[],\"Out\":[{\"Name\":\"Value\",\"Ident\":\"driver.Value\",\"Kind\":20,\"PkgPath\":\"database/sql/driver\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]}}}},\"immutable\":true,\"position\":{\"Index\":2,\"MixedIn\":false,\"MixinIndex\":0},\"schema_type\":{\"mysql\":\"decimal(18,10)\",\"postgres\":\"numeric(18,10)\"},\"annotations\":{\"EntGQL\":{\"Skip\":48,\"Type\":\"String\"}},\"comment\":\"The exchange rate from the native currency to user's default display currency\"}],\"annotations\":{\"EntGQL\":{\"QueryField\":{}},\"EntSQL\":{\"increment_start\":34359738368}}},{\"name\":\"Transaction\",\"config\":{\"Table\":\"\"},\"edges\":[{\"name\":\"profile\",\"type\":\"Profile\",\"ref_name\":\"transactions\",\"unique\":true,\"inverse\":true,\"required\":true,\"annotations\":{\"EntGQL\":{\"Skip\":48}}},{\"name\":\"category\",\"type\":\"Category\",\"ref_name\":\"transactions\",\"unique\":true,\"inverse\":true,\"required\":true,\"annotations\":{\"EntGQL\":{\"Skip\":48}}},{\"name\":\"transaction_entries\",\"type\":\"TransactionEntry\"}],\"fields\":[{\"name\":\"create_time\",\"type\":{\"Type\":2,\"Ident\":\"\",\"PkgPath\":\"time\",\"PkgName\":\"\",\"Nillable\":false,\"RType\":null},\"default\":true,\"default_kind\":19,\"immutable\":true,\"position\":{\"Index\":0,\"MixedIn\":true,\"MixinIndex\":0}},{\"name\":\"update_time\",\"type\":{\"Type\":2,\"Ident\":\"\",\"PkgPath\":\"time\",\"PkgName\":\"\",\"Nillable\":false,\"RType\":null},\"default\":true,\"default_kind\":19,\"update_default\":true,\"position\":{\"Index\":1,\"MixedIn\":true,\"MixinIndex\":0}},{\"name\":\"note\",\"type\":{\"Type\":7,\"Ident\":\"\",\"PkgPath\":\"\",\"PkgName\":\"\",\"Nillable\":false,\"RType\":null},\"size\":2147483647,\"optional\":true,\"position\":{\"Index\":0,\"MixedIn\":false,\"MixinIndex\":0}}],\"annotations\":{\"EntGQL\":{\"MutationInputs\":[{\"IsCreate\":true},{}],\"QueryField\":{},\"RelayConnection\":true},\"EntSQL\":{\"increment_start\":8589934592}}},{\"name\":\"TransactionEntry\",\"config\":{\"Table\":\"\"},\"edges\":[{\"name\":\"account\",\"type\":\"Account\",\"ref_name\":\"transaction_entries\",\"unique\":true,\"inverse\":true,\"required\":true},{\"name\":\"transaction\",\"type\":\"Transaction\",\"ref_name\":\"transaction_entries\",\"unique\":true,\"inverse\":true,\"required\":true}],\"fields\":[{\"name\":\"create_time\",\"type\":{\"Type\":2,\"Ident\":\"\",\"PkgPath\":\"time\",\"PkgName\":\"\",\"Nillable\":false,\"RType\":null},\"default\":true,\"default_kind\":19,\"immutable\":true,\"position\":{\"Index\":0,\"MixedIn\":true,\"MixinIndex\":0}},{\"name\":\"update_time\",\"type\":{\"Type\":2,\"Ident\":\"\",\"PkgPath\":\"time\",\"PkgName\":\"\",\"Nillable\":false,\"RType\":null},\"default\":true,\"default_kind\":19,\"update_default\":true,\"position\":{\"Index\":1,\"MixedIn\":true,\"MixinIndex\":0}},{\"name\":\"note\",\"type\":{\"Type\":7,\"Ident\":\"\",\"PkgPath\":\"\",\"PkgName\":\"\",\"Nillable\":false,\"RType\":null},\"size\":2147483647,\"optional\":true,\"position\":{\"Index\":0,\"MixedIn\":false,\"MixinIndex\":0}},{\"name\":\"amount\",\"type\":{\"Type\":20,\"Ident\":\"decimal.Decimal\",\"PkgPath\":\"github.com/shopspring/decimal\",\"PkgName\":\"decimal\",\"Nillable\":false,\"RType\":{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":{\"Abs\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Add\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Atan\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"BigFloat\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"*big.Float\",\"Kind\":22,\"PkgPath\":\"\",\"Methods\":null}]},\"BigInt\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"*big.Int\",\"Kind\":22,\"PkgPath\":\"\",\"Methods\":null}]},\"Ceil\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Cmp\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"int\",\"Ident\":\"int\",\"Kind\":2,\"PkgPath\":\"\",\"Methods\":null}]},\"Coefficient\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"*big.Int\",\"Kind\":22,\"PkgPath\":\"\",\"Methods\":null}]},\"CoefficientInt64\":{\"In\":[],\"Out\":[{\"Name\":\"int64\",\"Ident\":\"int64\",\"Kind\":6,\"PkgPath\":\"\",\"Methods\":null}]},\"Compare\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"int\",\"Ident\":\"int\",\"Kind\":2,\"PkgPath\":\"\",\"Methods\":null}]},\"Copy\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Cos\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Div\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"DivRound\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Equal\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"Equals\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"ExpHullAbrham\":{\"In\":[{\"Name\":\"uint32\",\"Ident\":\"uint32\",\"Kind\":10,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"ExpTaylor\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"Exponent\":{\"In\":[],\"Out\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}]},\"Float64\":{\"In\":[],\"Out\":[{\"Name\":\"float64\",\"Ident\":\"float64\",\"Kind\":14,\"PkgPath\":\"\",\"Methods\":null},{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"Floor\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"GobDecode\":{\"In\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"GobEncode\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"GreaterThan\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"GreaterThanOrEqual\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"InexactFloat64\":{\"In\":[],\"Out\":[{\"Name\":\"float64\",\"Ident\":\"float64\",\"Kind\":14,\"PkgPath\":\"\",\"Methods\":null}]},\"IntPart\":{\"In\":[],\"Out\":[{\"Name\":\"int64\",\"Ident\":\"int64\",\"Kind\":6,\"PkgPath\":\"\",\"Methods\":null}]},\"IsInteger\":{\"In\":[],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"IsNegative\":{\"In\":[],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"IsPositive\":{\"In\":[],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"IsZero\":{\"In\":[],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"LessThan\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"LessThanOrEqual\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"Ln\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"MarshalBinary\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"MarshalJSON\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"MarshalText\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"Mod\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Mul\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Neg\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"NumDigits\":{\"In\":[],\"Out\":[{\"Name\":\"int\",\"Ident\":\"int\",\"Kind\":2,\"PkgPath\":\"\",\"Methods\":null}]},\"Pow\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"PowBigInt\":{\"In\":[{\"Name\":\"\",\"Ident\":\"*big.Int\",\"Kind\":22,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"PowInt32\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"PowWithPrecision\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"QuoRem\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Rat\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"*big.Rat\",\"Kind\":22,\"PkgPath\":\"\",\"Methods\":null}]},\"Round\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"RoundBank\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"RoundCash\":{\"In\":[{\"Name\":\"uint8\",\"Ident\":\"uint8\",\"Kind\":8,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"RoundCeil\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"RoundDown\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"RoundFloor\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"RoundUp\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Scan\":{\"In\":[{\"Name\":\"\",\"Ident\":\"interface {}\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"Shift\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Sign\":{\"In\":[],\"Out\":[{\"Name\":\"int\",\"Ident\":\"int\",\"Kind\":2,\"PkgPath\":\"\",\"Methods\":null}]},\"Sin\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"String\":{\"In\":[],\"Out\":[{\"Name\":\"string\",\"Ident\":\"string\",\"Kind\":24,\"PkgPath\":\"\",\"Methods\":null}]},\"StringFixed\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"string\",\"Ident\":\"string\",\"Kind\":24,\"PkgPath\":\"\",\"Methods\":null}]},\"StringFixedBank\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"string\",\"Ident\":\"string\",\"Kind\":24,\"PkgPath\":\"\",\"Methods\":null}]},\"StringFixedCash\":{\"In\":[{\"Name\":\"uint8\",\"Ident\":\"uint8\",\"Kind\":8,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"string\",\"Ident\":\"string\",\"Kind\":24,\"PkgPath\":\"\",\"Methods\":null}]},\"StringScaled\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"string\",\"Ident\":\"string\",\"Kind\":24,\"PkgPath\":\"\",\"Methods\":null}]},\"Sub\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Tan\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Truncate\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"UnmarshalBinary\":{\"In\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"UnmarshalJSON\":{\"In\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"UnmarshalText\":{\"In\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"Value\":{\"In\":[],\"Out\":[{\"Name\":\"Value\",\"Ident\":\"driver.Value\",\"Kind\":20,\"PkgPath\":\"database/sql/driver\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]}}}},\"position\":{\"Index\":1,\"MixedIn\":false,\"MixinIndex\":0},\"schema_type\":{\"mysql\":\"decimal(36,18)\",\"postgres\":\"numeric(36,18)\"},\"annotations\":{\"EntGQL\":{\"Type\":\"String\"}},\"comment\":\"The unit amount of share or money in this transaction entry\"},{\"name\":\"value\",\"type\":{\"Type\":20,\"Ident\":\"decimal.Decimal\",\"PkgPath\":\"github.com/shopspring/decimal\",\"PkgName\":\"decimal\",\"Nillable\":false,\"RType\":{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":{\"Abs\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Add\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Atan\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"BigFloat\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"*big.Float\",\"Kind\":22,\"PkgPath\":\"\",\"Methods\":null}]},\"BigInt\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"*big.Int\",\"Kind\":22,\"PkgPath\":\"\",\"Methods\":null}]},\"Ceil\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Cmp\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"int\",\"Ident\":\"int\",\"Kind\":2,\"PkgPath\":\"\",\"Methods\":null}]},\"Coefficient\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"*big.Int\",\"Kind\":22,\"PkgPath\":\"\",\"Methods\":null}]},\"CoefficientInt64\":{\"In\":[],\"Out\":[{\"Name\":\"int64\",\"Ident\":\"int64\",\"Kind\":6,\"PkgPath\":\"\",\"Methods\":null}]},\"Compare\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"int\",\"Ident\":\"int\",\"Kind\":2,\"PkgPath\":\"\",\"Methods\":null}]},\"Copy\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Cos\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Div\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"DivRound\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Equal\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"Equals\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"ExpHullAbrham\":{\"In\":[{\"Name\":\"uint32\",\"Ident\":\"uint32\",\"Kind\":10,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"ExpTaylor\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"Exponent\":{\"In\":[],\"Out\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}]},\"Float64\":{\"In\":[],\"Out\":[{\"Name\":\"float64\",\"Ident\":\"float64\",\"Kind\":14,\"PkgPath\":\"\",\"Methods\":null},{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"Floor\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"GobDecode\":{\"In\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"GobEncode\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"GreaterThan\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"GreaterThanOrEqual\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"InexactFloat64\":{\"In\":[],\"Out\":[{\"Name\":\"float64\",\"Ident\":\"float64\",\"Kind\":14,\"PkgPath\":\"\",\"Methods\":null}]},\"IntPart\":{\"In\":[],\"Out\":[{\"Name\":\"int64\",\"Ident\":\"int64\",\"Kind\":6,\"PkgPath\":\"\",\"Methods\":null}]},\"IsInteger\":{\"In\":[],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"IsNegative\":{\"In\":[],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"IsPositive\":{\"In\":[],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"IsZero\":{\"In\":[],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"LessThan\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"LessThanOrEqual\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"Ln\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"MarshalBinary\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"MarshalJSON\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"MarshalText\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"Mod\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Mul\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Neg\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"NumDigits\":{\"In\":[],\"Out\":[{\"Name\":\"int\",\"Ident\":\"int\",\"Kind\":2,\"PkgPath\":\"\",\"Methods\":null}]},\"Pow\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"PowBigInt\":{\"In\":[{\"Name\":\"\",\"Ident\":\"*big.Int\",\"Kind\":22,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"PowInt32\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"PowWithPrecision\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"QuoRem\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Rat\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"*big.Rat\",\"Kind\":22,\"PkgPath\":\"\",\"Methods\":null}]},\"Round\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"RoundBank\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"RoundCash\":{\"In\":[{\"Name\":\"uint8\",\"Ident\":\"uint8\",\"Kind\":8,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"RoundCeil\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"RoundDown\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"RoundFloor\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"RoundUp\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Scan\":{\"In\":[{\"Name\":\"\",\"Ident\":\"interface {}\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"Shift\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Sign\":{\"In\":[],\"Out\":[{\"Name\":\"int\",\"Ident\":\"int\",\"Kind\":2,\"PkgPath\":\"\",\"Methods\":null}]},\"Sin\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"String\":{\"In\":[],\"Out\":[{\"Name\":\"string\",\"Ident\":\"string\",\"Kind\":24,\"PkgPath\":\"\",\"Methods\":null}]},\"StringFixed\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"string\",\"Ident\":\"string\",\"Kind\":24,\"PkgPath\":\"\",\"Methods\":null}]},\"StringFixedBank\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"string\",\"Ident\":\"string\",\"Kind\":24,\"PkgPath\":\"\",\"Methods\":null}]},\"StringFixedCash\":{\"In\":[{\"Name\":\"uint8\",\"Ident\":\"uint8\",\"Kind\":8,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"string\",\"Ident\":\"string\",\"Kind\":24,\"PkgPath\":\"\",\"Methods\":null}]},\"StringScaled\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"string\",\"Ident\":\"string\",\"Kind\":24,\"PkgPath\":\"\",\"Methods\":null}]},\"Sub\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Tan\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Truncate\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"UnmarshalBinary\":{\"In\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"UnmarshalJSON\":{\"In\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"UnmarshalText\":{\"In\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"Value\":{\"In\":[],\"Out\":[{\"Name\":\"Value\",\"Ident\":\"driver.Value\",\"Kind\":20,\"PkgPath\":\"database/sql/driver\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]}}}},\"position\":{\"Index\":2,\"MixedIn\":false,\"MixinIndex\":0},\"schema_type\":{\"mysql\":\"decimal(18,10)\",\"postgres\":\"numeric(18,10)\"},\"annotations\":{\"EntGQL\":{\"Skip\":48,\"Type\":\"String\"}},\"comment\":\"The value of 1 share in the native currency. If this is just a currency account, then this field will be 1\"},{\"name\":\"balance\",\"type\":{\"Type\":20,\"Ident\":\"decimal.Decimal\",\"PkgPath\":\"github.com/shopspring/decimal\",\"PkgName\":\"decimal\",\"Nillable\":false,\"RType\":{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":{\"Abs\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Add\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Atan\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"BigFloat\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"*big.Float\",\"Kind\":22,\"PkgPath\":\"\",\"Methods\":null}]},\"BigInt\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"*big.Int\",\"Kind\":22,\"PkgPath\":\"\",\"Methods\":null}]},\"Ceil\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Cmp\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"int\",\"Ident\":\"int\",\"Kind\":2,\"PkgPath\":\"\",\"Methods\":null}]},\"Coefficient\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"*big.Int\",\"Kind\":22,\"PkgPath\":\"\",\"Methods\":null}]},\"CoefficientInt64\":{\"In\":[],\"Out\":[{\"Name\":\"int64\",\"Ident\":\"int64\",\"Kind\":6,\"PkgPath\":\"\",\"Methods\":null}]},\"Compare\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"int\",\"Ident\":\"int\",\"Kind\":2,\"PkgPath\":\"\",\"Methods\":null}]},\"Copy\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Cos\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Div\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"DivRound\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Equal\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"Equals\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"ExpHullAbrham\":{\"In\":[{\"Name\":\"uint32\",\"Ident\":\"uint32\",\"Kind\":10,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"ExpTaylor\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"Exponent\":{\"In\":[],\"Out\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}]},\"Float64\":{\"In\":[],\"Out\":[{\"Name\":\"float64\",\"Ident\":\"float64\",\"Kind\":14,\"PkgPath\":\"\",\"Methods\":null},{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"Floor\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"GobDecode\":{\"In\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"GobEncode\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"GreaterThan\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"GreaterThanOrEqual\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"InexactFloat64\":{\"In\":[],\"Out\":[{\"Name\":\"float64\",\"Ident\":\"float64\",\"Kind\":14,\"PkgPath\":\"\",\"Methods\":null}]},\"IntPart\":{\"In\":[],\"Out\":[{\"Name\":\"int64\",\"Ident\":\"int64\",\"Kind\":6,\"PkgPath\":\"\",\"Methods\":null}]},\"IsInteger\":{\"In\":[],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"IsNegative\":{\"In\":[],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"IsPositive\":{\"In\":[],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"IsZero\":{\"In\":[],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"LessThan\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"LessThanOrEqual\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"bool\",\"Ident\":\"bool\",\"Kind\":1,\"PkgPath\":\"\",\"Methods\":null}]},\"Ln\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"MarshalBinary\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"MarshalJSON\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"MarshalText\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"Mod\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Mul\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Neg\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"NumDigits\":{\"In\":[],\"Out\":[{\"Name\":\"int\",\"Ident\":\"int\",\"Kind\":2,\"PkgPath\":\"\",\"Methods\":null}]},\"Pow\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"PowBigInt\":{\"In\":[{\"Name\":\"\",\"Ident\":\"*big.Int\",\"Kind\":22,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"PowInt32\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"PowWithPrecision\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"QuoRem\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null},{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Rat\":{\"In\":[],\"Out\":[{\"Name\":\"\",\"Ident\":\"*big.Rat\",\"Kind\":22,\"PkgPath\":\"\",\"Methods\":null}]},\"Round\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"RoundBank\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"RoundCash\":{\"In\":[{\"Name\":\"uint8\",\"Ident\":\"uint8\",\"Kind\":8,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"RoundCeil\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"RoundDown\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"RoundFloor\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"RoundUp\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Scan\":{\"In\":[{\"Name\":\"\",\"Ident\":\"interface {}\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"Shift\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Sign\":{\"In\":[],\"Out\":[{\"Name\":\"int\",\"Ident\":\"int\",\"Kind\":2,\"PkgPath\":\"\",\"Methods\":null}]},\"Sin\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"String\":{\"In\":[],\"Out\":[{\"Name\":\"string\",\"Ident\":\"string\",\"Kind\":24,\"PkgPath\":\"\",\"Methods\":null}]},\"StringFixed\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"string\",\"Ident\":\"string\",\"Kind\":24,\"PkgPath\":\"\",\"Methods\":null}]},\"StringFixedBank\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"string\",\"Ident\":\"string\",\"Kind\":24,\"PkgPath\":\"\",\"Methods\":null}]},\"StringFixedCash\":{\"In\":[{\"Name\":\"uint8\",\"Ident\":\"uint8\",\"Kind\":8,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"string\",\"Ident\":\"string\",\"Kind\":24,\"PkgPath\":\"\",\"Methods\":null}]},\"StringScaled\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"string\",\"Ident\":\"string\",\"Kind\":24,\"PkgPath\":\"\",\"Methods\":null}]},\"Sub\":{\"In\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Tan\":{\"In\":[],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"Truncate\":{\"In\":[{\"Name\":\"int32\",\"Ident\":\"int32\",\"Kind\":5,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"Decimal\",\"Ident\":\"decimal.Decimal\",\"Kind\":25,\"PkgPath\":\"github.com/shopspring/decimal\",\"Methods\":null}]},\"UnmarshalBinary\":{\"In\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"UnmarshalJSON\":{\"In\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"UnmarshalText\":{\"In\":[{\"Name\":\"\",\"Ident\":\"[]uint8\",\"Kind\":23,\"PkgPath\":\"\",\"Methods\":null}],\"Out\":[{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]},\"Value\":{\"In\":[],\"Out\":[{\"Name\":\"Value\",\"Ident\":\"driver.Value\",\"Kind\":20,\"PkgPath\":\"database/sql/driver\",\"Methods\":null},{\"Name\":\"error\",\"Ident\":\"error\",\"Kind\":20,\"PkgPath\":\"\",\"Methods\":null}]}}}},\"position\":{\"Index\":3,\"MixedIn\":false,\"MixinIndex\":0},\"schema_type\":{\"mysql\":\"decimal(36,18)\",\"postgres\":\"numeric(36,18)\"},\"annotations\":{\"EntGQL\":{\"Skip\":48,\"Type\":\"String\"}},\"comment\":\"The total balance of this transaction entry in this account's currency\"}],\"annotations\":{\"EntGQL\":{\"MutationInputs\":[{\"IsCreate\":true},{}]},\"EntSQL\":{\"increment_start\":12884901888}}},{\"name\":\"User\",\"config\":{\"Table\":\"\"},\"edges\":[{\"name\":\"user_keys\",\"type\":\"UserKey\",\"annotations\":{\"EntSQL\":{\"on_delete\":\"CASCADE\"}}},{\"name\":\"profiles\",\"type\":\"Profile\",\"through\":{\"N\":\"user_profiles\",\"T\":\"UserProfile\"}}],\"fields\":[{\"name\":\"create_time\",\"type\":{\"Type\":2,\"Ident\":\"\",\"PkgPath\":\"time\",\"PkgName\":\"\",\"Nillable\":false,\"RType\":null},\"default\":true,\"default_kind\":19,\"immutable\":true,\"position\":{\"Index\":0,\"MixedIn\":true,\"MixinIndex\":0}},{\"name\":\"update_time\",\"type\":{\"Type\":2,\"Ident\":\"\",\"PkgPath\":\"time\",\"PkgName\":\"\",\"Nillable\":false,\"RType\":null},\"default\":true,\"default_kind\":19,\"update_default\":true,\"position\":{\"Index\":1,\"MixedIn\":true,\"MixinIndex\":0}},{\"name\":\"email\",\"type\":{\"Type\":7,\"Ident\":\"\",\"PkgPath\":\"\",\"PkgName\":\"\",\"Nillable\":false,\"RType\":null},\"unique\":true,\"validators\":1,\"position\":{\"Index\":0,\"MixedIn\":false,\"MixinIndex\":0}}],\"annotations\":{\"EntGQL\":{\"MutationInputs\":[{\"IsCreate\":true}]},\"EntSQL\":{\"increment_start\":17179869184}}},{\"name\":\"UserKey\",\"config\":{\"Table\":\"\"},\"edges\":[{\"name\":\"user\",\"type\":\"User\",\"ref_name\":\"user_keys\",\"unique\":true,\"inverse\":true,\"required\":true}],\"fields\":[{\"name\":\"create_time\",\"type\":{\"Type\":2,\"Ident\":\"\",\"PkgPath\":\"time\",\"PkgName\":\"\",\"Nillable\":false,\"RType\":null},\"default\":true,\"default_kind\":19,\"immutable\":true,\"position\":{\"Index\":0,\"MixedIn\":true,\"MixinIndex\":0}},{\"name\":\"update_time\",\"type\":{\"Type\":2,\"Ident\":\"\",\"PkgPath\":\"time\",\"PkgName\":\"\",\"Nillable\":false,\"RType\":null},\"default\":true,\"default_kind\":19,\"update_default\":true,\"position\":{\"Index\":1,\"MixedIn\":true,\"MixinIndex\":0}},{\"name\":\"key\",\"type\":{\"Type\":7,\"Ident\":\"\",\"PkgPath\":\"\",\"PkgName\":\"\",\"Nillable\":false,\"RType\":null},\"validators\":1,\"position\":{\"Index\":0,\"MixedIn\":false,\"MixinIndex\":0}},{\"name\":\"hashed_password\",\"type\":{\"Type\":7,\"Ident\":\"\",\"PkgPath\":\"\",\"PkgName\":\"\",\"Nillable\":false,\"RType\":null},\"optional\":true,\"position\":{\"Index\":1,\"MixedIn\":false,\"MixinIndex\":0}}],\"annotations\":{\"EntGQL\":{\"MutationInputs\":[{\"IsCreate\":true}]},\"EntSQL\":{\"increment_start\":21474836480}}},{\"name\":\"UserProfile\",\"config\":{\"Table\":\"\"},\"edges\":[{\"name\":\"user\",\"type\":\"User\",\"field\":\"user_id\",\"unique\":true,\"required\":true},{\"name\":\"profile\",\"type\":\"Profile\",\"field\":\"profile_id\",\"unique\":true,\"required\":true}],\"fields\":[{\"name\":\"permission\",\"type\":{\"Type\":6,\"Ident\":\"userprofile.Permission\",\"PkgPath\":\"\",\"PkgName\":\"\",\"Nillable\":false,\"RType\":null},\"enums\":[{\"N\":\"owner\",\"V\":\"owner\"},{\"N\":\"admin\",\"V\":\"admin\"},{\"N\":\"viewer\",\"V\":\"viewer\"}],\"position\":{\"Index\":0,\"MixedIn\":false,\"MixinIndex\":0}},{\"name\":\"user_id\",\"type\":{\"Type\":12,\"Ident\":\"\",\"PkgPath\":\"\",\"PkgName\":\"\",\"Nillable\":false,\"RType\":null},\"position\":{\"Index\":1,\"MixedIn\":false,\"MixinIndex\":0}},{\"name\":\"profile_id\",\"type\":{\"Type\":12,\"Ident\":\"\",\"PkgPath\":\"\",\"PkgName\":\"\",\"Nillable\":false,\"RType\":null},\"position\":{\"Index\":2,\"MixedIn\":false,\"MixinIndex\":0}}],\"annotations\":{\"EntSQL\":{\"increment_start\":42949672960},\"Fields\":{\"ID\":[\"user_id\",\"profile_id\"],\"StructTag\":null}}}],\"Features\":[\"sql/globalid\",\"privacy\",\"schema/snapshot\",\"namedges\"]}"
