+++ apps/server/ent/mutation.go
@@ -2104,12 +2104,6 @@ func (m ProfileMutation) Tx() (*Tx, error) {
 	return tx, nil
 }
 
-// SetID sets the value of the id field. Note that this
-// operation is only accepted on creation of Profile entities.
-func (m *ProfileMutation) SetID(id int) {
-	m.id = &id
-}
-
 // ID returns the ID value in the mutation. Note that the ID is only available
 // if it was provided to the builder or after it was returned from the database.
 func (m *ProfileMutation) ID() (id int, exists bool) {
